Crate sui_types

Modules

accumulator

balance

base_types

certificate_proof

chain_id

coin

collection_types

committee

crypto

dynamic_field

error

event

filter

gas

gas_coin

id

in_memory_storage

intent

message_envelope

messages

messages_checkpoint

move_package

object

query

signature_seed

A secret seed value, useful for deterministic private key and SuiAddress generation.

storage

sui_serde

sui_system_state

temporary_store

utils

Macros

exit_main

fp_bail

fp_ensure

Constants

MOVE_STDLIB_ADDRESS

0x1– account address where Move stdlib modules are stored Same as the ObjectID

MOVE_STDLIB_OBJECT_ID

SUI_FRAMEWORK_ADDRESS

0x2– account address where sui framework modules are stored Same as the ObjectID

SUI_FRAMEWORK_OBJECT_ID

SUI_SYSTEM_STATE_OBJECT_ID

0x5: hardcoded object ID for the singleton sui system state object.

SUI_SYSTEM_STATE_OBJECT_SHARED_VERSION

Functions

parse_sui_struct_tag

parse_sui_type_tag

sui_framework_address_concat_string