sui_types

Module accumulator_root

Structs

AccumulatorKey

Rust type for the Move type accumulator::Key used to derive the dynamic field id for the accumulator value.

AccumulatorObjId

New-type for ObjectIDs that are known to have been properly derived as a Balance accumulator field.

EventCommitment

EventStreamHead

Rust representation of the Move EventStreamHead struct from accumulator_settlement module. This represents the state of an authenticated event stream head stored on-chain.

U128

Enums

AccumulatorValue

Constants

ACCUMULATOR_METADATA_MODULE

ACCUMULATOR_ROOT_CREATE_FUNC

ACCUMULATOR_ROOT_MODULE

ACCUMULATOR_ROOT_SETTLEMENT_PROLOGUE_FUNC

ACCUMULATOR_ROOT_SETTLEMENT_SETTLE_EVENTS_FUNC

ACCUMULATOR_ROOT_SETTLE_U128_FUNC

ACCUMULATOR_SETTLEMENT_EVENT_STREAM_HEAD

ACCUMULATOR_SETTLEMENT_MODULE

Functions

build_event_merkle_root

get_accumulator_root_obj_initial_shared_version

stream_id_from_accumulator_event

Extract stream id from an accumulator event if it targets sui::accumulator_settlement::EventStreamHead

update_account_balance_for_testing