sui_types

Module base_types

Re-exports

pub use crate::committee::EpochId;

pub use crate::digests::ObjectDigest;

pub use crate::digests::TransactionDigest;

pub use crate::digests::TransactionEffectsDigest;

Structs

ExecutionData

ExecutionDigests

FullObjectRef

MoveLegacyTxContext

MoveObjectType

Wrapper around StructTag with a space-efficient representation for common types like coins The StructTag for a gas coin is 84 bytes, so using 1 byte instead is a win. The inner representation is private to prevent incorrectly constructing an Other instead of one of the specialized variants, e.g. Other(GasCoin::type_()) instead of GasCoin

ObjectID

ObjectInfo

SequenceNumber

SizeOneVec

SuiAddress

TxContext

UserData

VerifiedExecutionData

Enums

FullObjectID

MoveObjectType_

Even though it is declared public, it is the “private”, internal representation for MoveObjectType

ObjectIDParseError

ObjectType

Type of a Sui object

TxContextKind

Constants

RESOLVED_ASCII_STR

RESOLVED_STD_OPTION

RESOLVED_TX_CONTEXT

RESOLVED_UTF8_STR

STD_ASCII_MODULE_NAME

STD_ASCII_STRUCT_NAME

STD_OPTION_MODULE_NAME

STD_OPTION_STRUCT_NAME

STD_UTF8_MODULE_NAME

STD_UTF8_STRUCT_NAME

SUI_ADDRESS_LENGTH

TX_CONTEXT_MODULE_NAME

TX_CONTEXT_STRUCT_NAME

URL_MODULE_NAME

URL_STRUCT_NAME

VEC_MAP_ENTRY_STRUCT_NAME

VEC_MAP_MODULE_NAME

VEC_MAP_STRUCT_NAME

Traits

ConciseableName

Functions

dbg_addr

Generate a fake SuiAddress with repeated one byte.

dbg_object_id

Generate a fake ObjectID with repeated one byte.

is_primitive_type_tag

Whether this type is valid as a primitive (pure) transaction input.

move_ascii_str_layout

move_utf8_str_layout

random_object_ref

update_object_ref_for_testing

url_layout

Type Aliases

AuthorityName

ConsensusObjectSequenceKey

Represents an distinct stream of object versions for a consensus object, based on the object ID and start version.

ObjectRef

TxSequenceNumber

VersionDigest

VersionNumber