sui_types
Module coin_registry

Structs

Currency

Currency stores metadata such as name, symbol, decimals, icon_url and description, as well as supply states (optional) and regulatory status.

CurrencyKey

Empty struct used as a key for deriving Currency addresses

Supply

TypeName

Enums

CurrencyRegulatedState

MetadataCapState

SupplyState

Constants

COIN_REGISTRY_MODULE_NAME

CURRENCY_KEY_STRUCT_NAME

Functions

get_coin_registry_obj_initial_shared_version

Type Aliases

ExtraField