Trait sui_types::crypto::SigningKey

pub trait SigningKey: 'static + ToFromBytes + Serialize + DeserializeOwned + Send + Sync {
    type PubKey: VerifyingKey<PrivKey = Self>;
    type Sig: Authenticator<PrivKey = Self>;

    const LENGTH: usize;
}

Trait impl’d by private (secret) keys in asymmetric cryptography.

The trait bounds are implemented so as to be symmetric and equivalent to the ones on its associated types for public key and signature material.

Required Associated Types

type PubKey: VerifyingKey<PrivKey = Self>

type Sig: Authenticator<PrivKey = Self>

Required Associated Constants

const LENGTH: usize