sui_types
Module deny_list_v1

Structs

CoinDenyCap

DenyList

Rust representation of the Move type 0x2::deny_list::DenyList. It has a bag that contains the deny lists for different system types. At creation, there is only one type (at key 0), which is the Coin type. We also take advantage of the dynamic nature of Bag to add more types in the future, as well as making changes to the deny lists for existing types.

PerTypeDenyList

Rust representation of the Move type 0x2::deny_list::PerTypeDenyList. denied_count is a table that stores the number of denied addresses for each coin template type. It can be used as a quick check to see if an address is denied for any coin template type. denied_addresses is a table that stores all the addresses that are denied for each coin template type. The key to the table is the coin template type in string form: package_id::module_name::struct_name.

RegulatedCoinMetadata

Constants

DENY_LIST_COIN_TYPE_INDEX

DENY_LIST_CREATE_FUNC

DENY_LIST_MODULE

Functions

check_coin_deny_list_v1

Checks coin denylist v1 at signing time. It checks that none of the coin types in the transaction are denied for the sender.

get_coin_deny_list

get_deny_list_obj_initial_shared_version

get_deny_list_root_object