sui_types::dynamic_field
Function get_dynamic_field_object_from_store

pub fn get_dynamic_field_object_from_store<K>(
    object_store: &dyn ObjectStore,
    parent_id: ObjectID,
    key: &K,
) -> Result<Object, SuiError>
where
    K: MoveTypeTagTrait + Serialize + DeserializeOwned + Debug,

Given a parent object ID (e.g. a table), and a key, retrieve the corresponding dynamic field object from the object_store. The key type K must implement MoveTypeTagTrait which has an associated function that returns the Move type tag. Note that this function returns the Field object itself, not the value in the field.