sui_types::effects
Enum TransactionEffects

pub enum TransactionEffects {
    V1(TransactionEffectsV1),
    V2(TransactionEffectsV2),
}

The response from processing a transaction or a certified transaction

Variants

V1(TransactionEffectsV1)

V2(TransactionEffectsV2)

Implementations

impl TransactionEffects

pub fn new_from_execution_v1( status: ExecutionStatus, executed_epoch: EpochId, gas_used: GasCostSummary, modified_at_versions: Vec<(ObjectID, SequenceNumber)>, shared_objects: Vec<ObjectRef>, transaction_digest: TransactionDigest, created: Vec<(ObjectRef, Owner)>, mutated: Vec<(ObjectRef, Owner)>, unwrapped: Vec<(ObjectRef, Owner)>, deleted: Vec<ObjectRef>, unwrapped_then_deleted: Vec<ObjectRef>, wrapped: Vec<ObjectRef>, gas_object: (ObjectRef, Owner), events_digest: Option<TransactionEventsDigest>, dependencies: Vec<TransactionDigest>, ) -> Self

Creates a TransactionEffects message from the results of execution, choosing the correct format for the current protocol version.

pub fn new_from_execution_v2( status: ExecutionStatus, executed_epoch: EpochId, gas_used: GasCostSummary, shared_objects: Vec<SharedInput>, loaded_per_epoch_config_objects: BTreeSet<ObjectID>, transaction_digest: TransactionDigest, lamport_version: SequenceNumber, changed_objects: BTreeMap<ObjectID, EffectsObjectChange>, gas_object: Option<ObjectID>, events_digest: Option<TransactionEventsDigest>, dependencies: Vec<TransactionDigest>, ) -> Self

Creates a TransactionEffects message from the results of execution, choosing the correct format for the current protocol version.

pub fn execution_digests(&self) -> ExecutionDigests

pub fn estimate_effects_size_upperbound_v1( num_writes: usize, num_mutables: usize, num_deletes: usize, num_deps: usize, ) -> usize

pub fn estimate_effects_size_upperbound_v2( num_writes: usize, num_modifies: usize, num_deps: usize, ) -> usize

pub fn all_changed_objects(&self) -> Vec<(ObjectRef, Owner, WriteKind)>

Return an iterator that iterates through all changed objects, including mutated, created and unwrapped objects. In other words, all objects that still exist in the object state after this transaction. It doesn’t include deleted/wrapped objects.

pub fn all_removed_objects(&self) -> Vec<(ObjectRef, ObjectRemoveKind)>

Return all objects that existed in the state prior to the transaction but no longer exist in the state after the transaction. It includes deleted and wrapped objects, but does not include unwrapped_then_deleted objects.

pub fn all_tombstones(&self) -> Vec<(ObjectID, SequenceNumber)>

Returns all objects that will become a tombstone after this transaction. This includes deleted, unwrapped_then_deleted and wrapped objects.

pub fn mutated_excluding_gas(&self) -> Vec<(ObjectRef, Owner)>

Return an iterator of mutated objects, but excluding the gas object.

pub fn summary_for_debug(&self) -> TransactionEffectsDebugSummary