sui_types::error

Enum SuiErrorKind

pub enum SuiErrorKind {

    UserInputError {
        error: UserInputError,
    },
    SuiObjectResponseError {
        error: SuiObjectResponseError,
    },
    UnexpectedOwnerType,
    TooManyTransactionsPendingExecution {
        queue_len: usize,
        threshold: usize,
    },
    TooManyTransactionsPendingConsensus,
    TooManyTransactionsPendingOnObject {
        object_id: ObjectID,
        queue_len: usize,
        threshold: usize,
    },
    TooOldTransactionPendingOnObject {
        object_id: ObjectID,
        txn_age_sec: u64,
        threshold: u64,
    },
    InvalidTxKindInSoftBundle,
    InvalidSignature {
        error: String,
    },
    SignerSignatureAbsent {
        expected: String,
        actual: Vec<String>,
    },
    SignerSignatureNumberMismatch {
        expected: usize,
        actual: usize,
    },
    IncorrectSigner {
        error: String,
    },
    UnknownSigner {
        signer: Option<String>,
        index: Option<u32>,
        committee: Box<Committee>,
    },
    StakeAggregatorRepeatedSigner {
        signer: AuthorityName,
        conflicting_sig: bool,
    },
    PotentiallyTemporarilyInvalidSignature {
        error: String,
    },
    WrongEpoch {
        expected_epoch: EpochId,
        actual_epoch: EpochId,
    },
    CertificateRequiresQuorum,
    DEPRECATED_ErrorWhileProcessingCertificate,
    QuorumFailedToGetEffectsQuorumWhenProcessingTransaction {
        effects_map: BTreeMap<TransactionEffectsDigest, (Vec<AuthorityName>, StakeUnit)>,
    },
    FailedToVerifyTxCertWithExecutedEffects {
        validator_name: AuthorityName,
        error: String,
    },
    TxAlreadyFinalizedWithDifferentUserSigs,
    InvalidAuthenticator,
    InvalidAddress,
    InvalidTransactionDigest,
    InvalidDigestLength {
        expected: usize,
        actual: usize,
    },
    InvalidDkgMessageSize,
    UnexpectedMessage(String),
    ModuleVerificationFailure {
        error: String,
    },
    ModuleDeserializationFailure {
        error: String,
    },
    ModulePublishFailure {
        error: String,
    },
    ModuleBuildFailure {
        error: String,
    },
    FunctionNotFound {
        error: String,
    },
    ModuleNotFound {
        module_name: String,
    },
    TypeError {
        error: String,
    },
    CircularObjectOwnership,
    ObjectLockAlreadyInitialized {
        refs: Vec<ObjectRef>,
    },
    ObjectLockConflict {
        obj_ref: ObjectRef,
        pending_transaction: TransactionDigest,
    },
    ObjectLockedAtFutureEpoch {
        obj_refs: Vec<ObjectRef>,
        locked_epoch: EpochId,
        new_epoch: EpochId,
        locked_by_tx: TransactionDigest,
    },
    TransactionNotFound {
        digest: TransactionDigest,
    },
    TransactionsNotFound {
        digests: Vec<TransactionDigest>,
    },
    TransactionEventsNotFound {
        digest: TransactionDigest,
    },
    TransactionEffectsNotFound {
        digest: TransactionDigest,
    },
    TransactionAlreadyExecuted {
        digest: TransactionDigest,
    },
    TransactionRejectReasonNotFound {
        digest: TransactionDigest,
    },
    BadObjectType {
        error: String,
    },
    FailObjectLayout {
        st: String,
    },
    ExecutionInvariantViolation,
    ByzantineAuthoritySuspicion {
        authority: AuthorityName,
        reason: String,
    },
    DEPRECATED_StorageError,
    DEPRECATED_GenericStorageError,
    InvalidChildObjectAccess {
        object: ObjectID,
        given_parent: ObjectID,
        actual_owner: Owner,
    },
    DEPRECATED_StorageMissingFieldError,
    DEPRECATED_StorageCorruptedFieldError,
    GenericAuthorityError {
        error: String,
    },
    GenericBridgeError {
        error: String,
    },
    FailedToDispatchSubscription {
        error: String,
    },
    OwnerFailedToSerialize {
        error: String,
    },
    ExtraFieldFailedToDeserialize {
        error: String,
    },
    TransactionOrchestratorLocalExecutionError {
        error: String,
    },
    TransactionSerializationError {
        error: String,
    },
    TransactionDeserializationError {
        error: String,
    },
    TransactionEffectsSerializationError {
        error: String,
    },
    TransactionEffectsDeserializationError {
        error: String,
    },
    TransactionEventsSerializationError {
        error: String,
    },
    TransactionEventsDeserializationError {
        error: String,
    },
    ObjectSerializationError {
        error: String,
    },
    ObjectDeserializationError {
        error: String,
    },
    NoEventStore,
    TooManyIncorrectAuthorities {
        errors: Vec<(AuthorityName, SuiError)>,
        action: String,
    },
    FullNodeInvalidTxRangeQuery {
        error: String,
    },
    FailedToSubmitToConsensus(String),
    ConsensusConnectionBroken(String),
    HandleConsensusTransactionFailure(String),
    SignatureKeyGenError(String),
    KeyConversionError(String),
    InvalidPrivateKey,
    FullNodeCantHandleCertificate,
    ValidatorHaltedAtEpochEnd,
    EpochEnded(EpochId),
    AdvanceEpochError {
        error: String,
    },
    TransactionExpired,
    RpcError(String, String),
    InvalidRpcMethodError,
    UnsupportedFeatureError {
        error: String,
    },
    QuorumDriverCommunicationError {
        error: String,
    },
    TimeoutError,
    ExecutionError(String),
    InvalidCommittee(String),
    MissingCommitteeAtEpoch(EpochId),
    IndexStoreNotAvailable,
    DynamicFieldReadError(String),
    SuiSystemStateReadError(String),
    SuiBridgeReadError(String),
    UnexpectedVersion(String),
    WrongMessageVersion {
        error: String,
    },
    Unknown(String),
    FileIOError(String),
    JWKRetrievalError,
    Storage(String),
    ValidatorOverloadedRetryAfter {
        retry_after_secs: u64,
    },
    TooManyRequests,
    NoCertificateProvidedError,
    NitroAttestationFailedToVerify(String),
    GrpcMessageSerializeError {
        type_info: String,
        error: String,
    },
    GrpcMessageDeserializeError {
        type_info: String,
        error: String,
    },
    ValidatorConsensusLagging {
        round: u32,
        last_committed_round: u32,
    },
    InvalidAdminRequest(String),
    InvalidRequest(String),
    AliasesChanged,
    ImmutableObjectClaimNotFoundInInput {
        object_id: ObjectID,
    },
    ImmutableObjectNotClaimed {
        object_id: ObjectID,
    },
    InvalidImmutableObjectClaim {
        claimed_object_id: ObjectID,
        found_object_ref: ObjectRef,
    },

}

Custom error type for Sui.

Variants

UserInputError

Fields

error: UserInputError

SuiObjectResponseError

Fields

error: SuiObjectResponseError

UnexpectedOwnerType

TooManyTransactionsPendingExecution

Fields

queue_len: usize

threshold: usize

TooManyTransactionsPendingConsensus

TooManyTransactionsPendingOnObject

Fields

object_id: ObjectID

queue_len: usize

threshold: usize

TooOldTransactionPendingOnObject

Fields

object_id: ObjectID

txn_age_sec: u64

threshold: u64

InvalidTxKindInSoftBundle

InvalidSignature

Fields

error: String

SignerSignatureAbsent

Fields

expected: String

actual: Vec<String>

SignerSignatureNumberMismatch

Fields

expected: usize

actual: usize

IncorrectSigner

Fields

error: String

UnknownSigner

Fields

signer: Option<String>

index: Option<u32>

committee: Box<Committee>

StakeAggregatorRepeatedSigner

Fields

signer: AuthorityName

conflicting_sig: bool

PotentiallyTemporarilyInvalidSignature

Fields

error: String

WrongEpoch

Fields

expected_epoch: EpochId

actual_epoch: EpochId

CertificateRequiresQuorum

DEPRECATED_ErrorWhileProcessingCertificate

QuorumFailedToGetEffectsQuorumWhenProcessingTransaction

Fields

effects_map: BTreeMap<TransactionEffectsDigest, (Vec<AuthorityName>, StakeUnit)>

FailedToVerifyTxCertWithExecutedEffects

Fields

validator_name: AuthorityName

error: String

TxAlreadyFinalizedWithDifferentUserSigs

InvalidAuthenticator

InvalidAddress

InvalidTransactionDigest

InvalidDigestLength

Fields

expected: usize

actual: usize

InvalidDkgMessageSize

UnexpectedMessage(String)

ModuleVerificationFailure

Fields

error: String

ModuleDeserializationFailure

Fields

error: String

ModulePublishFailure

Fields

error: String

ModuleBuildFailure

Fields

error: String

FunctionNotFound

Fields

error: String

ModuleNotFound

Fields

module_name: String

TypeError

Fields

error: String

CircularObjectOwnership

ObjectLockAlreadyInitialized

Fields

refs: Vec<ObjectRef>

ObjectLockConflict

Fields

obj_ref: ObjectRef

pending_transaction: TransactionDigest

ObjectLockedAtFutureEpoch

Fields

obj_refs: Vec<ObjectRef>

locked_epoch: EpochId

new_epoch: EpochId

locked_by_tx: TransactionDigest

TransactionNotFound

Fields

digest: TransactionDigest

TransactionsNotFound

Fields

digests: Vec<TransactionDigest>

TransactionEventsNotFound

Fields

digest: TransactionDigest

TransactionEffectsNotFound

Fields

digest: TransactionDigest

TransactionAlreadyExecuted

Fields

digest: TransactionDigest

TransactionRejectReasonNotFound

Fields

digest: TransactionDigest

BadObjectType

Fields

error: String

FailObjectLayout

Fields

st: String

ExecutionInvariantViolation

ByzantineAuthoritySuspicion

Fields

authority: AuthorityName

reason: String

DEPRECATED_StorageError

DEPRECATED_GenericStorageError

InvalidChildObjectAccess

Fields

object: ObjectID

given_parent: ObjectID

actual_owner: Owner

DEPRECATED_StorageMissingFieldError

DEPRECATED_StorageCorruptedFieldError

GenericAuthorityError

Fields

error: String

GenericBridgeError

Fields

error: String

FailedToDispatchSubscription

Fields

error: String

OwnerFailedToSerialize

Fields

error: String

ExtraFieldFailedToDeserialize

Fields

error: String

TransactionOrchestratorLocalExecutionError

Fields

error: String

TransactionSerializationError

Fields

error: String

TransactionDeserializationError

Fields

error: String

TransactionEffectsSerializationError

Fields

error: String

TransactionEffectsDeserializationError

Fields

error: String

TransactionEventsSerializationError

Fields

error: String

TransactionEventsDeserializationError

Fields

error: String

ObjectSerializationError

Fields

error: String

ObjectDeserializationError

Fields

error: String

NoEventStore

TooManyIncorrectAuthorities

Fields

errors: Vec<(AuthorityName, SuiError)>

action: String

FullNodeInvalidTxRangeQuery

Fields

error: String

FailedToSubmitToConsensus(String)

ConsensusConnectionBroken(String)

HandleConsensusTransactionFailure(String)

SignatureKeyGenError(String)

KeyConversionError(String)

InvalidPrivateKey

FullNodeCantHandleCertificate

ValidatorHaltedAtEpochEnd

EpochEnded(EpochId)

AdvanceEpochError

Fields

error: String

TransactionExpired

RpcError(String, String)

InvalidRpcMethodError

UnsupportedFeatureError

Fields

error: String

QuorumDriverCommunicationError

Fields

error: String

TimeoutError

ExecutionError(String)

InvalidCommittee(String)

MissingCommitteeAtEpoch(EpochId)

IndexStoreNotAvailable

DynamicFieldReadError(String)

SuiSystemStateReadError(String)

SuiBridgeReadError(String)

UnexpectedVersion(String)

WrongMessageVersion

Fields

error: String

Unknown(String)

FileIOError(String)

JWKRetrievalError

Storage(String)

ValidatorOverloadedRetryAfter

Fields

retry_after_secs: u64

TooManyRequests

NoCertificateProvidedError

NitroAttestationFailedToVerify(String)

GrpcMessageSerializeError

Fields

type_info: String

error: String

GrpcMessageDeserializeError

Fields

type_info: String

error: String

ValidatorConsensusLagging

Fields

round: u32

last_committed_round: u32

InvalidAdminRequest(String)

InvalidRequest(String)

AliasesChanged

ImmutableObjectClaimNotFoundInInput

Fields

object_id: ObjectID

ImmutableObjectNotClaimed

Fields

object_id: ObjectID

InvalidImmutableObjectClaim

Fields

claimed_object_id: ObjectID

found_object_ref: ObjectRef

Implementations

impl SuiErrorKind

pub fn to_variant_name(&self) -> &'static str

Returns the variant name of the error. Sub-variants within UserInputError are unpacked too.

pub fn individual_error_indicates_epoch_change(&self) -> bool

pub fn is_retryable(&self) -> (bool, bool)

Returns if the error is retryable and if the error’s retryability is explicitly categorized. There should be only a handful of retryable errors. For now we list common non-retryable error below to help us find more retryable errors in logs.

pub fn is_object_or_package_not_found(&self) -> bool

pub fn is_overload(&self) -> bool

pub fn is_retryable_overload(&self) -> bool

pub fn retry_after_secs(&self) -> u64

pub fn categorize(&self) -> ErrorCategory

Categorizes SuiError into ErrorCategory.