Enum sui_types::event::Event

pub enum Event {
    MoveEvent {
        package_id: ObjectID,
        transaction_module: Identifier,
        sender: SuiAddress,
        type_: StructTag,
        contents: Vec<u8>,
    },
    Publish {
        sender: SuiAddress,
        package_id: ObjectID,
        version: SequenceNumber,
        digest: ObjectDigest,
    },
    CoinBalanceChange {
        package_id: ObjectID,
        transaction_module: Identifier,
        sender: SuiAddress,
        change_type: BalanceChangeType,
        owner: Owner,
        coin_type: String,
        coin_object_id: ObjectID,
        version: SequenceNumber,
        amount: i128,
    },
    EpochChange(EpochId),
    Checkpoint(CheckpointSequenceNumber),
    TransferObject {
        package_id: ObjectID,
        transaction_module: Identifier,
        sender: SuiAddress,
        recipient: Owner,
        object_type: String,
        object_id: ObjectID,
        version: SequenceNumber,
    },
    MutateObject {
        package_id: ObjectID,
        transaction_module: Identifier,
        sender: SuiAddress,
        object_type: String,
        object_id: ObjectID,
        version: SequenceNumber,
    },
    DeleteObject {
        package_id: ObjectID,
        transaction_module: Identifier,
        sender: SuiAddress,
        object_id: ObjectID,
        version: SequenceNumber,
    },
    NewObject {
        package_id: ObjectID,
        transaction_module: Identifier,
        sender: SuiAddress,
        recipient: Owner,
        object_type: String,
        object_id: ObjectID,
        version: SequenceNumber,
    },
}

Specific type of event

Variants

MoveEvent

Fields

package_id: ObjectID

transaction_module: Identifier

sender: SuiAddress

type_: StructTag

contents: Vec<u8>

Transaction level event Move-specific event

Publish

Fields

sender: SuiAddress

package_id: ObjectID

version: SequenceNumber

digest: ObjectDigest

Module published

CoinBalanceChange

Fields

package_id: ObjectID

transaction_module: Identifier

sender: SuiAddress

change_type: BalanceChangeType

owner: Owner

coin_type: String

coin_object_id: ObjectID

version: SequenceNumber

amount: i128

The amount indicate the coin value changes for this event, negative amount means spending coin value and positive means receiving coin value.

Coin balance changing event

EpochChange(EpochId)

Epoch change

Checkpoint(CheckpointSequenceNumber)

New checkpoint

TransferObject

Fields

package_id: ObjectID

transaction_module: Identifier

sender: SuiAddress

recipient: Owner

object_type: String

object_id: ObjectID

version: SequenceNumber

Object level event Transfer objects to new address / wrap in another object

MutateObject

Fields

package_id: ObjectID

transaction_module: Identifier

sender: SuiAddress

object_type: String

object_id: ObjectID

version: SequenceNumber

Object level event Object mutated.

DeleteObject

Fields

package_id: ObjectID

transaction_module: Identifier

sender: SuiAddress

object_id: ObjectID

version: SequenceNumber

Delete object

NewObject

Fields

package_id: ObjectID

transaction_module: Identifier

sender: SuiAddress

recipient: Owner

object_type: String

object_id: ObjectID

version: SequenceNumber

New object creation

Implementations

impl Event

pub fn move_event(
package_id: &AccountAddress,
module: &IdentStr,
sender: SuiAddress,
type_: StructTag,
contents: Vec<u8>
) -> Self

pub fn balance_change(
ctx: &SingleTxContext,
change_type: BalanceChangeType,
owner: Owner,
coin_object_id: ObjectID,
version: SequenceNumber,
object_type: &StructTag,
amount: i128
) -> Self

pub fn transfer_object(
ctx: &SingleTxContext,
recipient: Owner,
object_type: String,
object_id: ObjectID,
version: SequenceNumber
) -> Self

pub fn new_object(
ctx: &SingleTxContext,
recipient: Owner,
object_type: String,
object_id: ObjectID,
version: SequenceNumber
) -> Self

pub fn name_from_ordinal(ordinal: usize) -> &'static str

pub fn event_type(&self) -> EventType

Returns the EventType associated with an Event

pub fn object_id(&self) -> Option<ObjectID>

Returns the object or package ID associated with the event, if available. Specifically:

For TransferObject: the object ID being transferred (eg moving child from parent, its the child)

for DeleteObject and NewObject, the Object ID

pub fn package_id(&self) -> Option<ObjectID>

Extracts the Move package ID associated with the event, or the package published.

pub fn sender(&self) -> Option<SuiAddress>

Extracts the Sender address associated with the event.

pub fn module_name(&self) -> Option<&str>

Extract a module name, if available, from a SuiEvent

pub fn recipient(&self) -> Option<&Owner>

Extracts the recipient from a SuiEvent, if available

pub fn recipient_serialized(&self) -> Result<Option<String>, SuiError>

Extracts the serialized recipient from a SuiEvent, if available

pub fn move_event_contents(&self) -> Option<&[u8]>

Extracts the bcs move content from a SuiEvent, if available

pub fn move_event_name(&self) -> Option<String>

Extracts the move event name (StructTag) from a SuiEvent, if available “0x2::devnet_nft::MintNFTEvent”

pub fn object_type(&self) -> Option<String>

Extracts the TransferType from a SuiEvent, if available

pub fn object_version(&self) -> Option<&SequenceNumber>

Extracts the Object Version from a SuiEvent, if available

pub fn digest(&self) -> Option<ObjectDigest>

Extracts the digest from a SuiEvent::Publish event

pub fn amount(&self) -> Option<i128>

Extracts the amount from a SuiEvent::CoinBalanceChange event

pub fn balance_change_type(&self) -> Option<&BalanceChangeType>

Extracts the balance change type from a SuiEvent::CoinBalanceChange event

pub fn balance_change_from_ordinal(
ordinal: usize
) -> Result<BalanceChangeType, SuiError>

pub fn move_event_to_move_struct(
type_: &StructTag,
contents: &[u8],
resolver: &impl GetModule
) -> Result<MoveStruct, SuiError>