Module sui_types::execution

Structs

DynamicallyLoadedObjectMetadata

ExecutionResultsV1

ExecutionResultsV2

Used by sui-execution v1 and above, to capture the execution results from Move. The results represent the primitive information that can then be used to construct both transaction effects V1 and V2.

InputValue

ObjectValue

ResultValue

Enums

CommandKind

ExecutionResults

InputObjectMetadata

ObjectContents

RawValueType

SharedInput

UsageKind

Value

Traits

ExecutionState

Interface with the store necessary to execute a programmable transaction

SuiResolver

TryFromValue

TypeLayoutStore

View of the store necessary to produce the layouts of types.

Functions

get_denied_certificates

If a transaction digest shows up in this list, when executing such transaction, we will always return ExecutionError::CertificateDenied without executing it (but still do gas smashing). Because this list is not gated by protocol version, there are a few important criteria for adding a digest to this list:

is_certificate_denied

Type Aliases

DeletedSharedObjectInfo

A type containing all of the information needed to work with a deleted shared object in execution and when committing the execution effects of the transaction. This holds: 0. The object ID of the deleted shared object.

DeletedSharedObjects

A sequence of information about deleted shared objects in the transaction’s inputs.