sui_types::execution
Type Alias ConsensusStreamEndedInfo

pub type ConsensusStreamEndedInfo = (ObjectID, SequenceNumber, bool, TransactionDigest);

A type containing all of the information needed to work in execution with an object whose consensus stream is ended, and when committing the execution effects of the transaction. This holds: 0. The object ID.

The version.

Whether the object appeared as mutable (or owned) in the transaction, or as read-only.

The transaction digest of the previous transaction that used this object mutably or took it by value.