sui_types::execution_params

Function get_early_execution_error

pub fn get_early_execution_error(
    transaction_digest: &TransactionDigest,
    input_objects: &CheckedInputObjects,
    config_certificate_deny_set: &HashSet<TransactionDigest>,
    funds_withdraw_status: &FundsWithdrawStatus,
) -> Option<ExecutionErrorKind>

Determine if a transaction is predetermined to fail execution. If so, return the error kind, otherwise return None. When we pass this to the execution engine, we will not execute the transaction if it is predetermined to fail execution.