Module sui_types::gas

Structs

GasCostSummary

SuiGasStatus

Statics

MAX_GAS_BUDGET

MIN_GAS_BUDGET

Functions

check_gas_balance

Check whether the given gas_object and gas_budget is legit:

deduct_gas

Subtract the gas balance of \p gas_object by \p amount. This function should never fail, since we checked that the budget is always less than balance, and the amount is capped at the budget.

get_gas_balance

refund_gas

start_gas_metering

Create a new gas status with the given gas_budget, and charge the transaction flat fee.