Function sui_types::gas::check_gas_balance

pub fn check_gas_balance(
    gas_object: &Object,
    gas_budget: u64,
    gas_price: u64,
    more_gas_objs: Vec<Object>,
    cost_table: &SuiCostTable
) -> UserInputResult

Check whether the given gas_object and gas_budget is legit:

If the gas object has an address owner.

If it’s enough to pay the flat minimum transaction fee

If it’s less than the max gas budget allowed

If the gas_object actually has enough balance to pay for the budget

If total balance in gas object and extra input objects is sufficient to pay total amount of gas budget and extra amount to pay, extra input objects and extra amount to pay are only relevant in SUI payment transactions.