sui_types::gas_model

Module tables

Structs

GasStatus

The Move VM implementation of state for gas metering.

Constants

CONST_SIZE

The size in bytes for a non-string or address constant on the stack

MIN_EXISTS_DATA_SIZE

For exists checks on data that doesn’t exists this is the multiplier that is used.

REFERENCE_SIZE

The size in bytes for a reference on the stack

STRUCT_SIZE

The size of a struct in bytes

VEC_SIZE

The size of a vector (without its containing data) in bytes

VM_FLAT_FEE

VM flat fee

Statics

INITIAL_COST_SCHEDULE

ZERO_COST_SCHEDULE

Functions

initial_cost_schedule_for_unit_tests

initial_cost_schedule_v1

initial_cost_schedule_v2

initial_cost_schedule_v3

initial_cost_schedule_v4

initial_cost_schedule_v5

unit_cost_schedule

zero_cost_schedule