sui_types

Module governance

Structs

StakedSui

Constants

ADD_STAKE_FUN_NAME

ADD_STAKE_MUL_COIN_FUN_NAME

MAX_VALIDATOR_COUNT

Maximum number of active validators at any moment. We do not allow the number of validators in any epoch to go above this.

MIN_VALIDATOR_JOINING_STAKE_MISTDeprecated

Lower-bound on the amount of stake required to become a validator.

SIP_39_PHASE_LENGTH

Number of epochs for a single phase of SIP-39 since the change

STAKED_SUI_STRUCT_NAME

STAKING_POOL_MODULE_NAME

VALIDATOR_LOW_POWER_PHASE_1

Low voting power threshold for validators in Phase 1. Validators below this threshold fall into the “at risk” group. .08% of voting power

VALIDATOR_LOW_POWER_PHASE_2

Low voting power threshold for validators in Phase 2. Validators below this threshold fall into the “at risk” group. .08% of voting power

VALIDATOR_LOW_POWER_PHASE_3

Low voting power threshold for validators in Phase 3. Validators below this threshold fall into the “at risk” group. .02% of voting power

VALIDATOR_LOW_STAKE_GRACE_PERIOD

A validator can have stake below validator_low_stake_threshold for this many epochs before being kicked out.

VALIDATOR_LOW_STAKE_THRESHOLD_MISTDeprecated

Deprecated: with SIP-39 there is no longer a minimum stake requirement.

VALIDATOR_MIN_POWER_PHASE_1

Minimum amount of voting power required to become a validator in Phase 1. .12% of voting power

VALIDATOR_MIN_POWER_PHASE_2

Minimum amount of voting power required to become a validator in Phase 2. .12% of voting power

VALIDATOR_MIN_POWER_PHASE_3

Minimum amount of voting power required to become a validator in Phase 3. .03% of voting power

VALIDATOR_VERY_LOW_POWER_PHASE_1

Very low voting power threshold for validators in Phase 1. Validators below this threshold will be removed immediately at epoch change. .04% of voting power

VALIDATOR_VERY_LOW_POWER_PHASE_2

Very low voting power threshold for validators in Phase 2. Validators below this threshold will be removed immediately at epoch change. .04% of voting power

VALIDATOR_VERY_LOW_POWER_PHASE_3

Very low voting power threshold for validators in Phase 3. Validators below this threshold will be removed immediately at epoch change. .01% of voting power

VALIDATOR_VERY_LOW_STAKE_THRESHOLD_MISTDeprecated

Validators with stake below validator_very_low_stake_threshold will be removed immediately at epoch change, no grace period.

WITHDRAW_STAKE_FUN_NAME