Module sui_types::message_envelope

Structs

Envelope

TrustedEnvelope

TrustedEnvelope is a serializable wrapper around Envelope which is Into<VerifiedEnvelope> - in other words it models a verified message which has been written to the db (or some other trusted store), and may be read back from the db without further signature verification.

VerifiedEnvelope

Statics

FACEBOOK_JWK_BYTES

GOOGLE_JWK_BYTES

TWITCH_JWK_BYTES

Traits

AuthenticatedMessage

A message type that has an internal authenticator, such as SenderSignedData

Message

UnauthenticatedMessage

A marker trait to indicate !AuthenticatedMessage since rust does not allow negative trait bounds.

Functions

get_facebook_jwk_bytes

get_google_jwk_bytes

get_twitch_jwk_bytes