Trait sui_types::message_envelope::AuthenticatedMessage

pub trait AuthenticatedMessage {
    // Required methods
    fn verify_message_signature(
        &self,
        verify_params: &VerifyParams
    ) -> SuiResult;
    fn verify_uncached_checks(&self, verify_params: &VerifyParams) -> SuiResult;
}

A message type that has an internal authenticator, such as SenderSignedData

Required Methods

fn verify_message_signature(&self, verify_params: &VerifyParams) -> SuiResult

Verify internal signatures, e.g. for Transaction we verify the user signature(s).

fn verify_uncached_checks(&self, verify_params: &VerifyParams) -> SuiResult

Checks that still need to be verified outside cache.