Trait sui_types::message_envelope::Message

pub trait Message {
    type DigestType: Clone + Debug;

    const SCOPE: IntentScope;

    // Required methods
    fn digest(&self) -> Self::DigestType;
    fn verify_user_input(&self) -> SuiResult;
    fn verify_epoch(&self, epoch: EpochId) -> SuiResult;

    // Provided method
    fn scope(&self) -> IntentScope { ... }
}

Required Associated Types

type DigestType: Clone + Debug

Required Associated Constants

const SCOPE: IntentScope

Required Methods

fn digest(&self) -> Self::DigestType

fn verify_user_input(&self) -> SuiResult

Perform cheap validity checks before any expensive crypto verification.

fn verify_epoch(&self, epoch: EpochId) -> SuiResult

Verify that the message is from the correct epoch (e.g. for CertifiedCheckpointSummary we verify that the checkpoint is from the same epoch as the committee signatures).

Provided Methods

fn scope(&self) -> IntentScope

Object Safety

This trait is not object safe.