Struct sui_types::messages::TransactionEnvelope

pub struct TransactionEnvelope<S> {
    pub is_verified: bool,
    pub data: TransactionData,
    pub tx_signature: Signature,
    pub auth_sign_info: S,
    /* private fields */
}

A transaction signed by a client, optionally signed by an authority (depending on S). S indicates the authority signing state. It can be either empty or signed. We make the authority signature templated so that TransactionEnvelope<S> can be used universally in the transactions storage in SuiDataStore, shared by both authorities and non-authorities: authorities store signed transactions, while non-authorities store unsigned transactions.

Fields
is_verified: booldata: TransactionDatatx_signature: Signature

tx_signature is signed by the transaction sender, applied on data.

auth_sign_info: S

authority signature information, if available, is signed by an authority, applied on data.

Implementations

impl<S> TransactionEnvelope<S>

pub fn verify_signature(&self) -> Result<(), SuiError>

pub fn add_tx_sig_to_verification_obligation(
&self,
obligation: &mut VerificationObligation
) -> SuiResult<()>

pub fn sender_address(&self) -> SuiAddress

pub fn gas_payment_object_ref(&self) -> &ObjectRef

pub fn contains_shared_object(&self) -> bool

pub fn shared_input_objects(&self) -> impl Iterator<Item = &ObjectID>

pub fn digest(&self) -> &TransactionDigest

Get the transaction digest and write it to the cache

pub fn input_objects_in_compiled_modules(
compiled_modules: &[CompiledModule]
) -> Vec<InputObjectKind>

impl TransactionEnvelope<EmptySignInfo>

pub fn new(data: TransactionData, signature: Signature) -> Self

impl TransactionEnvelope<AuthoritySignInfo>

pub fn new(
epoch: EpochId,
transaction: Transaction,
authority: AuthorityName,
secret: &dyn Signer<AuthoritySignature>
) -> Self

Use signing key to create a signed object.

pub fn verify(&self, committee: &Committee) -> Result<usize, SuiError>

Verify the signature and return the non-zero voting right of the authority.

pub fn to_transaction(self) -> Transaction

impl TransactionEnvelope<AuthorityQuorumSignInfo>

pub fn new(transaction: Transaction) -> CertifiedTransaction

pub fn new_with_signatures(
epoch: EpochId,
transaction: Transaction,
signatures: Vec<(AuthorityName, AuthoritySignature)>
) -> CertifiedTransaction

pub fn to_transaction(self) -> Transaction

pub fn verify(&self, committee: &Committee) -> Result<(), SuiError>

Verify the certificate.

pub fn add_to_verification_obligation(
&self,
committee: &Committee,
obligation: &mut VerificationObligation
) -> SuiResult<()>