sui_types

Module messages_checkpoint

Re-exports

pub use crate::digests::CheckpointContentsDigest;

pub use crate::digests::CheckpointDigest;

Structs

CheckpointArtifacts

CheckpointContentsV1

CheckpointContents are the transactions included in an upcoming checkpoint. They must have already been causally ordered. Since the causal order algorithm is the same among validators, we expect all honest validators to come up with the same order for each checkpoint content.

CheckpointContentsV2

CheckpointRequest

CheckpointRequestV2

CheckpointResponse

CheckpointResponseV2

CheckpointSignatureMessage

This is a message validators publish to consensus in order to sign checkpoint

CheckpointSummary

CheckpointTransactionContents

CheckpointVersionSpecificDataV1

ECMHLiveObjectSetDigest

The Sha256 digest of an EllipticCurveMultisetHash committing to the live object set.

EndOfEpochData

FullCheckpointContents

Deprecated version of full checkpoint contents corresponding to CheckpointContentsV1.

FullCheckpointContentsV2

VerifiedCheckpointContents

Enums

CheckpointArtifact

CheckpointArtifact is a type that represents various artifacts of a checkpoint. We hash all the artifacts together to get the checkpoint artifacts digest that is included in the checkpoint summary.

CheckpointCommitment

CheckpointContents

CheckpointContentsView

CheckpointSummaryResponse

CheckpointVersionSpecificData

Holds data in CheckpointSummary that is serialized into the version_specific_data field.

VerifiedUserSignatures

VersionedFullCheckpointContents

Same as CheckpointContents, but contains full contents of all transactions, effects, and user signatures associated with the checkpoint.

Type Aliases

CertifiedCheckpointSummary

CheckpointSequenceNumber

CheckpointSummaryEnvelope

CheckpointTimestamp

SignedCheckpointSummary

TrustedCheckpoint

VerifiedCheckpoint