Type Definition sui_types::messages_checkpoint::CertifiedCheckpointSummary

pub type CertifiedCheckpointSummary = CheckpointSummaryEnvelope<AuthorityWeakQuorumSignInfo>;

Implementations

impl CertifiedCheckpointSummary

pub fn aggregate(
signed_checkpoints: Vec<SignedCheckpointSummary>,
committee: &Committee
) -> Result<CertifiedCheckpointSummary, SuiError>

Aggregate many checkpoint signatures to form a checkpoint certificate.

pub fn signatory_authorities<'a>(
&'a self,
committee: &'a Committee
) -> impl Iterator<Item = SuiResult<&AuthorityName>>

pub fn verify(
&self,
committee: &Committee,
contents: Option<&CheckpointContents>
) -> Result<(), SuiError>

Check that a certificate is valid, and signed by a quorum of authorities