Type Definition sui_types::messages_checkpoint::SignedCheckpointSummary

pub type SignedCheckpointSummary = CheckpointSummaryEnvelope<AuthoritySignInfo>;

Implementations

impl SignedCheckpointSummary

pub fn new(
epoch: EpochId,
sequence_number: CheckpointSequenceNumber,
network_total_transactions: u64,
authority: AuthorityName,
signer: &dyn Signer<AuthoritySignature>,
transactions: &CheckpointContents,
previous_digest: Option<CheckpointDigest>,
epoch_rolling_gas_cost_summary: GasCostSummary,
next_epoch_committee: Option<Committee>
) -> SignedCheckpointSummary

Create a new signed checkpoint proposal for this authority

pub fn new_from_summary(
checkpoint: CheckpointSummary,
authority: AuthorityName,
signer: &dyn Signer<AuthoritySignature>
) -> SignedCheckpointSummary

pub fn authority(&self) -> &AuthorityName

pub fn verify(
&self,
committee: &Committee,
contents: Option<&CheckpointContents>
) -> Result<(), SuiError>

Checks that the signature on the digest is correct, and verify the contents as well if provided.