sui_types

Module messages_grpc

Structs

ExecutedData

HandleCertificateRequestV3

HandleCertificateResponseV2

HandleCertificateResponseV3

Response type for version 3 of the handle certificate validator API.

HandleSoftBundleCertificatesRequestV3

Soft Bundle request. See SIP-19.

HandleSoftBundleCertificatesResponseV3

Response type for the handle Soft Bundle certificates validator API. If wait_for_effects is true, it is guaranteed that:

HandleTransactionResponse

ObjectInfoRequest

A request for information about an object and optionally its parent certificate at a specific version.

ObjectInfoResponse

This message provides information about the latest object and its lock.

RawExecutedData

RawExecutedStatus

RawExpiredStatus

RawRejectedStatus

RawSubmitTxRequest

RawSubmitTxResponse

RawSubmitTxResult

RawValidatorHealthRequest

Raw protobuf request for validator health information (evolvable)

RawValidatorHealthResponse

Raw protobuf response with validator health metrics (evolvable)

RawWaitForEffectsRequest

RawWaitForEffectsResponse

SubmitCertificateResponse

SubmitTxRequest

Contains either a transaction, or the type of Ping request.

SubmitTxResponse

SystemStateRequest

TransactionInfoRequest

TransactionInfoResponse

TxTypeIter

An iterator over the variants of TxType

ValidatorHealthRequest

Request for validator health information (used for latency measurement)

ValidatorHealthResponse

Response with validator health metrics (data collected but not used for scoring yet)

VerifiedHandleCertificateResponse

VerifiedObjectInfoResponse

Verified version of ObjectInfoResponse. layout and lock_for_debugging are skipped because they are not needed and we don’t want to verify them.

WaitForEffectsRequest

Enums

LayoutGenerationOption

Layout generation options – you can either generate or not generate the layout.

ObjectInfoRequestKind

PingType

RawValidatorSubmitStatus

RawValidatorTransactionStatus

SubmitTxResult

SubmitTxType

TransactionStatus

TxType

WaitForEffectsResponse