sui_types::object
Module option_visitor

Shared OptionVisitor implementation for deserializing Move Option types.

Structs

OptionVisitor

A visitor that deserializes an Option<T> by interpreting an empty vector as None and a single-element vector as Some(T).

Traits

OptionVisitorError

Trait for visitor errors that can be used with OptionVisitor. Requires the ability to create an “unexpected type” error.