sui_types::object

Module rpc_visitor

Modules

proto

Structs

LocalMeter

A meter backed by mutable local budget state.

RpcVisitor

A visitor that serializes Move values into some representation appropriate for RPC outputs.

Unmetered

Enums

Error

MeterError

Traits

Format

A trait for serializing values into some nested structured representation that supports null, bool, numbers, strings, vectors, and maps (e.g. JSON or Protobuf).

Meter

A trait for tracking the resource limits consumed while serializing a value.