Struct sui_types::object::Object

pub struct Object {
    pub data: Data,
    pub owner: Owner,
    pub previous_transaction: TransactionDigest,
    pub storage_rebate: u64,
}

Fields
data: Data

The meat of the object

owner: Owner

The owner that unlocks this object

previous_transaction: TransactionDigest

The digest of the transaction that created or last mutated this object

storage_rebate: u64

The amount of SUI we would rebate if this object gets deleted. This number is re-calculated each time the object is mutated based on the present storage gas price.

Implementations

impl Object

pub fn new_move(
o: MoveObject,
owner: Owner,
previous_transaction: TransactionDigest
) -> Self

Create a new Move object

pub fn new_package(
modules: Vec<CompiledModule>,
previous_transaction: TransactionDigest
) -> Result<Self, ExecutionError>

pub fn is_immutable(&self) -> bool

pub fn is_address_owned(&self) -> bool

pub fn is_child_object(&self) -> bool

pub fn is_shared(&self) -> bool

pub fn get_single_owner(&self) -> Option<SuiAddress>

pub fn get_owner_and_id(&self) -> Option<(Owner, ObjectID)>

pub fn is_package(&self) -> bool

Return true if this object is a Move package, false if it is a Move value

pub fn compute_object_reference(&self) -> ObjectRef

pub fn id(&self) -> ObjectID

pub fn version(&self) -> SequenceNumber

pub fn type_(&self) -> Option<&StructTag>

pub fn digest(&self) -> ObjectDigest

pub fn input_object_kind(&self) -> InputObjectKind

pub fn object_size_for_gas_metering(&self) -> usize

Approximate size of the object in bytes. This is used for gas metering. This will be slgihtly different from the serialized size, but we also don’t want to serialize the object just to get the size. This approximation should be good enough for gas metering.

pub fn transfer(&mut self, new_owner: SuiAddress)

Change the owner of self to new_owner.

pub fn get_layout(
&self,
format: ObjectFormatOptions,
resolver: &impl GetModule
) -> Result<Option<MoveStructLayout>, SuiError>

Get a MoveStructLayout for self. The resolver value must contain the module that declares self.type_ and the (transitive) dependencies of self.type_ in order for this to succeed. Failure will result in an ObjectSerializationError

pub fn get_move_template_type(&self) -> SuiResult<TypeTag>

Treat the object type as a Move struct with one type parameter, like this: S<T>. Returns the inner parameter type T.

pub fn ensure_public_transfer_eligible(&self) -> Result<(), ExecutionError>

impl Object

pub fn immutable_with_id_for_testing(id: ObjectID) -> Self

pub fn shared_for_testing() -> Object

make a test shared object.

pub fn with_id_owner_gas_for_testing(
id: ObjectID,
owner: SuiAddress,
gas: u64
) -> Self

pub fn with_object_owner_for_testing(id: ObjectID, owner: ObjectID) -> Self

pub fn with_id_owner_for_testing(id: ObjectID, owner: SuiAddress) -> Self

pub fn with_id_owner_version_for_testing(
id: ObjectID,
version: SequenceNumber,
owner: SuiAddress
) -> Self

pub fn with_owner_for_testing(owner: SuiAddress) -> Self

pub fn new_gas_with_balance_and_owner_for_testing(
value: u64,
owner: SuiAddress
) -> Self

Generate a new gas coin worth value with a random object ID and owner For testing purposes only

pub fn new_gas_for_testing() -> Self

Generate a new gas coin object with default balance and random owner.