sui_types::object
Struct ObjectInner

pub struct ObjectInner {
    pub data: Data,
    pub owner: Owner,
    pub previous_transaction: TransactionDigest,
    pub storage_rebate: u64,
}

Fields
data: Data

The meat of the object

owner: Owner

The owner that unlocks this object

previous_transaction: TransactionDigest

The digest of the transaction that created or last mutated this object

storage_rebate: u64

The amount of SUI we would rebate if this object gets deleted. This number is re-calculated each time the object is mutated based on the present storage gas price.

Implementations

impl ObjectInner

pub fn is_system_package(&self) -> bool

Returns true if the object is a system package.

pub fn is_immutable(&self) -> bool

pub fn is_address_owned(&self) -> bool

pub fn is_child_object(&self) -> bool

pub fn is_shared(&self) -> bool

pub fn is_consensus(&self) -> bool

pub fn get_single_owner(&self) -> Option<SuiAddress>

pub fn get_owner_and_id(&self) -> Option<(Owner, ObjectID)>

pub fn is_package(&self) -> bool

Return true if this object is a Move package, false if it is a Move value

pub fn compute_object_reference(&self) -> ObjectRef

pub fn compute_full_object_reference(&self) -> FullObjectRef

pub fn digest(&self) -> ObjectDigest

pub fn id(&self) -> ObjectID

pub fn full_id(&self) -> FullObjectID

pub fn version(&self) -> SequenceNumber

pub fn type_(&self) -> Option<&MoveObjectType>

pub fn struct_tag(&self) -> Option<StructTag>

pub fn is_coin(&self) -> bool

pub fn is_gas_coin(&self) -> bool

pub fn as_coin_maybe(&self) -> Option<Coin>

pub fn coin_type_maybe(&self) -> Option<TypeTag>

pub fn get_coin_value_unsafe(&self) -> u64

Return the value: u64 field of a Coin<T> type. Useful for reading the coin without deserializing the object into a Move value It is the caller’s responsibility to check that self is a coin–this function may panic or do something unexpected otherwise.

pub fn object_size_for_gas_metering(&self) -> usize

Approximate size of the object in bytes. This is used for gas metering. This will be slightly different from the serialized size, but we also don’t want to serialize the object just to get the size. This approximation should be good enough for gas metering.

pub fn transfer(&mut self, new_owner: SuiAddress)

Change the owner of self to new_owner.

pub fn get_layout( &self, resolver: &impl GetModule, ) -> Result<Option<MoveStructLayout>, SuiError>

Get a MoveStructLayout for self. The resolver value must contain the module that declares self.type_ and the (transitive) dependencies of self.type_ in order for this to succeed. Failure will result in an ObjectSerializationError

pub fn get_move_template_type(&self) -> SuiResult<TypeTag>

Treat the object type as a Move struct with one type parameter, like this: S<T>. Returns the inner parameter type T.

pub fn to_rust<'de, T: Deserialize<'de>>(&'de self) -> Option<T>