Struct sui_types::signature_seed::SignatureSeed

pub struct SignatureSeed(_);

A secret seed required for various cryptographic purposes, i.e., deterministic key derivation.

Instances of this seed are automatically overwritten with zeroes when they fall out of scope.

Implementations

impl SignatureSeed

pub fn to_bytes(&self) -> [u8; 32]

Convert this seed value to a byte array.

pub fn as_bytes(&self) -> &[u8; 32]

View this seed as a byte array.

pub fn from_bytes(bytes: &[u8]) -> Result<SignatureSeed, SuiError>

Construct a Seed from a slice of bytes.

Example

use sui_types::signature_seed::SignatureSeed;
use sui_types::error::SuiError;
use sui_types::signature_seed::SEED_LENGTH;
let secret_bytes: [u8; SEED_LENGTH] = [
   112, 012, 187, 211, 011, 092, 030, 001,
   225, 255, 000, 166, 112, 236, 044, 196,
   068, 073, 197, 105, 123, 050, 105, 025,
   112, 059, 172, 003, 028, 174, 127, 096, ];

let seed: SignatureSeed = SignatureSeed::from_bytes(&secret_bytes)?;
Ok(seed)

Input

A byte array that represents the secret bytes of this SignatureSeed.

Returns

A Result whose okay value is a SignatureSeed or whose error value is a SuiError wrapping the internal error that occurred.

pub fn generate<T>(csprng: &mut T) -> SignatureSeed where
T: CryptoRng + RngCore,

Generate a SignatureSeed from a csprng.

Example

extern crate rand;
use rand::rngs::OsRng;
use sui_types::signature_seed::SignatureSeed;
    let mut csprng = OsRng{};
    let secret_key: SignatureSeed = SignatureSeed::generate(&mut csprng);

Input

A CSPRNG with a fill_bytes() method, e.g. rand::OsRng

Returns

A fresh random SignatureSeed.

pub fn new_deterministic_address(
&self,
id: &[u8],
domain: Option<&[u8]>
) -> Result<SuiAddress, SuiError>

Deterministically generate a SuiAddress via HKDF.

Example

use serde::{Deserialize, Serialize};
use sui_types::signature_seed::SignatureSeed;

    // In production this SHOULD be a secret seed value, here we pin it for demo purposes.
    let seed = SignatureSeed::from_bytes(&[5u8; 32]).unwrap();

    // An input id.
    let id = "some-user@some-domain.com".as_bytes();

    // Some domain.
    let domain = "some-application".as_bytes();

    // Get address for the provided `id`.
    let sui_address = seed
        .new_deterministic_address(&id, Some(&domain))
        .unwrap();

Input

A user id byte-array, i.e., a username or email address. A domain separation value (optional), to distinguish between purposes of key derivation.

Returns

A derived SuiAddress, generated deterministically from some seed, id and domain.

pub fn sign<T>(
&self,
id: &[u8],
domain: Option<&[u8]>,
value: &T
) -> Result<Signature, Error> where
T: Signable<Vec<u8>>,

Sign a message using a deterministically derived key from some id input.

Example

use serde::{Deserialize, Serialize};
use sui_types::crypto::BcsSignable;
use sui_types::signature_seed::SignatureSeed;

#[derive(Serialize, Deserialize)]
struct Foo(String);

impl BcsSignable for Foo {}

    // In production this SHOULD be a secret seed value, here we pin it for demo purposes.
    let seed = SignatureSeed::from_bytes(&[5u8; 32]).unwrap();

    // An input id.
    let id = "some-user@some-domain.com".as_bytes();

    // Some domain.
    let domain = "some-application".as_bytes();

    // The msg to sign (note that we can only sign `Signable` objects.
    let msg = Foo("some-signable-message".to_string());

    let signature = seed.sign(&id, Some(&domain), &msg).unwrap();

    // Get address for the provided `id`.
    let sui_address = seed
        .new_deterministic_address(&id, Some(&domain))
        .unwrap();
    let verification = signature.verify(&msg, sui_address);
    assert!(verification.is_ok());

Input

A user id byte-array, i.e., a username or email address. A domain separation value (optional), to distinguish between purposes of key derivation. A Signable value (the message to be signed).

Returns

A Result whose okay value is a Signature or whose error value is a signature::Error wrapping the internal error that occurred.