Trait sui_types::storage::BackingPackageStore

pub trait BackingPackageStore {
    // Required method
    fn get_package_object(
        &self,
        package_id: &ObjectID
    ) -> SuiResult<Option<Object>>;

    // Provided methods
    fn get_package(
        &self,
        package_id: &ObjectID
    ) -> SuiResult<Option<MovePackage>> { ... }
    fn get_package_objects<'a>(
        &self,
        package_ids: impl IntoIterator<Item = &'a ObjectID>
    ) -> SuiResult<PackageFetchResults<Object>> { ... }
    fn get_packages<'a>(
        &self,
        package_ids: impl IntoIterator<Item = &'a ObjectID>
    ) -> SuiResult<PackageFetchResults<MovePackage>> { ... }
}

Required Methods

fn get_package_object(&self, package_id: &ObjectID) -> SuiResult<Option<Object>>

Provided Methods

fn get_package(&self, package_id: &ObjectID) -> SuiResult<Option<MovePackage>>

fn get_package_objects<'a>( &self, package_ids: impl IntoIterator<Item = &'a ObjectID> ) -> SuiResult<PackageFetchResults<Object>>

Returns Ok(<object for each package id in package_ids>) if all package IDs in package_id were found. If any package in package_ids was not found it returns a list of any package ids that are unable to be found>).

fn get_packages<'a>( &self, package_ids: impl IntoIterator<Item = &'a ObjectID> ) -> SuiResult<PackageFetchResults<MovePackage>>