sui_types

Module sui_sdk_types_conversions

Module for conversions between sui-core types and sui-sdk types

For now this module makes heavy use of the bcs_convert_impl macro to implement the From trait for converting between core and external sdk types, relying on the fact that the BCS format of these types are strictly identical. As time goes on we’ll slowly hand implement these impls directly to avoid going through the BCS machinery.

Structs

SdkTypeConversionError

Functions

struct_tag_core_to_sdk

struct_tag_sdk_to_core

type_tag_core_to_sdk

type_tag_sdk_to_core