sui_types

Module sui_system_state

Modules

epoch_start_sui_system_state

mock

sui_system_state_inner_v1

sui_system_state_inner_v2

sui_system_state_summary

Structs

AdvanceEpochParams

PoolTokenExchangeRate

SuiSystemStateWrapper

Rust version of the Move sui::sui_system::SuiSystemState type This represents the object with 0x5 ID. In Rust, this type should be rarely used since it’s just a thin wrapper used to access the inner object. Within this module, we use it to determine the current version of the system state inner object type, so that we could deserialize the inner object correctly. Outside of this module, we only use it in genesis snapshot and testing.

ValidatorWrapper

Enums

SuiSystemState

SuiSystemState provides an abstraction over multiple versions of the inner SuiSystemStateInner object. This should be the primary interface to the system state object in Rust. We use enum dispatch to dispatch all methods defined in SuiSystemStateTrait to the actual implementation in the inner types.

Constants

ADVANCE_EPOCH_FUNCTION_NAME

ADVANCE_EPOCH_SAFE_MODE_FUNCTION_NAME

SUI_SYSTEM_MODULE_NAME

SUI_SYSTEM_STATE_INNER_MODULE_NAME

SUI_SYSTEM_STATE_INNER_V1_STRUCT_NAME

SUI_SYSTEM_STATE_INNER_V2_STRUCT_NAME

VALIDATOR_MODULE_NAME

VALIDATOR_STRUCT_NAME

Traits

SuiSystemStateTrait

This is the standard API that all inner system state object type should implement.

Functions

get_sui_system_state

get_sui_system_state_wrapper

get_validator_from_table

Given a system state type version, and the ID of the table, along with a key, retrieve the dynamic field as a Validator type. We need the version to determine which inner type to use for the Validator type. This is assuming that the validator is stored in the table as ValidatorWrapper type.

get_validators_from_table_vec

Type Aliases

SuiSystemStateInnerGenesis

This is the fixed type used by genesis.

SuiValidatorGenesis