Module sui_types::sui_system_state::epoch_start_sui_system_state

Structs

EpochStartSystemStateV1

EpochStartValidatorInfoV1

Enums

EpochStartSystemState

This type captures the minimum amount of information from SuiSystemState needed by a validator to run the protocol. This allows us to decouple from the actual SuiSystemState type, and hence do not need to evolve it when we upgrade the SuiSystemState type. Evolving EpochStartSystemState is also a lot easier in that we could add optional fields and fill them with None for older versions. When we absolutely must delete fields, we could also add new db tables to store the new version. This is OK because we only store one copy of this as part of EpochStartConfiguration for the most recent epoch in the db.

Traits

EpochStartSystemStateTrait