sui_types::sui_system_state::sui_system_state_summary
Function get_validator_by_pool_id

pub fn get_validator_by_pool_id<S>(
    object_store: &S,
    system_state: &SuiSystemState,
    system_state_summary: &SuiSystemStateSummary,
    pool_id: ObjectID,
) -> Result<SuiValidatorSummary, SuiError>
where
    S: ObjectStore + ?Sized,

Given the staking pool id of a validator, return the validator’s SuiValidatorSummary, works for validator candidates, active validators, as well as inactive validators.