sui_types
Module supported_protocol_versions

Structs

ProtocolConfig

Constants that change the behavior of the protocol.

ProtocolVersion

SupportedProtocolVersions

Models the set of protocol versions supported by a validator. The sui-node binary will always use the SYSTEM_DEFAULT constant, but for testing we need to be able to inject arbitrary versions into SuiNode.

SupportedProtocolVersionsWithHashes

Models the set of protocol versions supported by a validator. The sui-node binary will always use the SYSTEM_DEFAULT constant, but for testing we need to be able to inject arbitrary versions into SuiNode.

Enums

Chain