sui_types::test_checkpoint_data_builder
Struct TestCheckpointDataBuilder

pub struct TestCheckpointDataBuilder { /* private fields */ }

A builder for creating test checkpoint data. Once initialized, the builder can be used to build multiple checkpoints. Call start_transaction to begin creating a new transaction. Call finish_transaction to complete the current transaction and add it to the current checkpoint. After all transactions are added, call build_checkpoint to get the final checkpoint data. This will also increment the stored checkpoint sequence number. Start the above process again to build the next checkpoint. NOTE: The generated checkpoint data is not guaranteed to be semantically valid or consistent. For instance, all object digests will be randomly set. It focuses on providing a way to generate various shaped test data for testing purposes. If you need to test the validity of the checkpoint data, you should use Simulacrum instead.

Implementations

impl TestCheckpointDataBuilder

pub fn new(checkpoint: u64) -> Self

pub fn with_epoch(self, epoch: u64) -> Self

Set the epoch for the checkpoint.

pub fn with_network_total_transactions( self, network_total_transactions: u64, ) -> Self

Set the network_total_transactions for the checkpoint.

pub fn with_timestamp_ms(self, timestamp_ms: CheckpointTimestamp) -> Self

Set the timestamp for the checkpoint.

pub fn start_transaction(self, sender_idx: u8) -> Self

Start creating a new transaction. sender_idx is a convenient representation of the sender’s address. A proper SuiAddress will be derived from it. It will also create a gas object for the sender if it doesn’t already exist in the live object map. You do not need to create the gas object yourself.

pub fn create_owned_object(self, object_idx: u64) -> Self

Create a new object in the transaction. object_idx is a convenient representation of the object’s ID. The object will be created as a SUI coin object, with default balance, and the transaction sender as its owner.

pub fn create_shared_object(self, object_idx: u64) -> Self

Create a new shared object in the transaction. object_idx is a convenient representation of the object’s ID. The object will be created as a SUI coin object, with default balance, and it is a shared object.

pub fn create_sui_object(self, object_idx: u64, balance: u64) -> Self

Create a new SUI coin object in the transaction. object_idx is a convenient representation of the object’s ID. balance is the amount of SUI to be created.

pub fn create_coin_object( self, object_idx: u64, owner_idx: u8, balance: u64, coin_type: TypeTag, ) -> Self

Create a new coin object in the transaction. object_idx is a convenient representation of the object’s ID. owner_idx is a convenient representation of the object’s owner’s address. balance is the amount of SUI to be created. coin_type is the type of the coin to be created.

pub fn mutate_owned_object(self, object_idx: u64) -> Self

Mutate an existing owned object in the transaction. object_idx is a convenient representation of the object’s ID.

pub fn mutate_shared_object(self, object_idx: u64) -> Self

Mutate an existing shared object in the transaction.

pub fn transfer_object(self, object_idx: u64, recipient_idx: u8) -> Self

Transfer an existing object to a new owner. object_idx is a convenient representation of the object’s ID. recipient_idx is a convenient representation of the recipient’s address.

pub fn change_object_owner(self, object_idx: u64, owner: Owner) -> Self

Change the owner of an existing object. object_idx is a convenient representation of the object’s ID. owner is the new owner of the object.

pub fn transfer_coin_balance( self, object_idx: u64, new_object_idx: u64, recipient_idx: u8, amount: u64, ) -> Self

Transfer part of an existing coin object’s balance to a new owner. object_idx is a convenient representation of the object’s ID. new_object_idx is a convenient representation of the new object’s ID. recipient_idx is a convenient representation of the recipient’s address. amount is the amount of balance to be transferred.

pub fn wrap_object(self, object_idx: u64) -> Self

Wrap an existing object in the transaction. object_idx is a convenient representation of the object’s ID.

pub fn unwrap_object(self, object_idx: u64) -> Self

Unwrap an existing object from the transaction. object_idx is a convenient representation of the object’s ID.

pub fn delete_object(self, object_idx: u64) -> Self

Delete an existing object from the transaction. object_idx is a convenient representation of the object’s ID.

pub fn read_frozen_object(self, object_id: u64) -> Self

Add an immutable object as an input to the transaction.

Fails if the object is not live or if its owner is not Owner::Immutable).

pub fn read_shared_object(self, object_idx: u64) -> Self

Add a read to a shared object to the transaction’s effects.

pub fn with_events(self, events: Vec<Event>) -> Self

Add events to the transaction. events is a vector of events to be added to the transaction.

pub fn add_move_call( self, package: ObjectID, module: &'static str, function: &'static str, ) -> Self

Add a move call PTB command to the transaction. package is the ID of the package to be called. module is the name of the module to be called. function is the name of the function to be called.

pub fn finish_transaction(self) -> Self

Complete the current transaction and add it to the checkpoint. This will also finalize all the object changes, and reflect them in the live object map.

pub fn build_checkpoint(&mut self) -> CheckpointData

Build the checkpoint data using all the transactions added to the builder so far. This will also increment the stored checkpoint sequence number.

pub fn advance_epoch(&mut self, safe_mode: bool) -> CheckpointData

Like advance_epoch_and_protocol_upgrade, but default the protocol version to the maximum.

pub fn advance_epoch_and_protocol_upgrade( &mut self, safe_mode: bool, protocol_version: ProtocolVersion, ) -> CheckpointData

Creates a transaction that advances the epoch, adds it to the checkpoint, and then builds the checkpoint. This increments the stored checkpoint sequence number and epoch. If safe_mode is true, the epoch end transaction will not include the SystemEpochInfoEvent. The protocol_version is used to set the protocol that we are going to follow in the subsequent epoch.

pub fn derive_object_id(object_idx: u64) -> ObjectID

Derive an object ID from an index. This is used to conveniently represent an object’s ID. We ensure that the bytes of object IDs have a stable order that is the same as object_idx.

pub fn derive_address(address_idx: u8) -> SuiAddress

Derive an address from an index.