sui_types::transaction

Trait TransactionDataAPI

pub trait TransactionDataAPI {

    // Required methods
    fn sender(&self) -> SuiAddress;
    fn kind(&self) -> &TransactionKind;
    fn kind_mut(&mut self) -> &mut TransactionKind;
    fn into_kind(self) -> TransactionKind;
    fn required_signers(&self) -> NonEmpty<SuiAddress>;
    fn gas_data(&self) -> &GasData;
    fn gas_owner(&self) -> SuiAddress;
    fn gas(&self) -> &[ObjectRef] ;
    fn gas_price(&self) -> u64;
    fn gas_budget(&self) -> u64;
    fn expiration(&self) -> &TransactionExpiration;
    fn expiration_mut(&mut self) -> &mut TransactionExpiration;
    fn move_calls(&self) -> Vec<(usize, &ObjectID, &str, &str)>;
    fn input_objects(&self) -> UserInputResult<Vec<InputObjectKind>>;
    fn shared_input_objects(&self) -> Vec<SharedInputObject>;
    fn receiving_objects(&self) -> Vec<ObjectRef> ;
    fn fastpath_dependency_objects(
        &self,
    ) -> UserInputResult<(Vec<ObjectRef>, Vec<ObjectID>, Vec<ObjectRef>)>;
    fn process_funds_withdrawals_for_signing(
        &self,
        chain_identifier: ChainIdentifier,
        coin_resolver: &dyn CoinReservationResolverTrait,
    ) -> UserInputResult<BTreeMap<AccumulatorObjId, u64>>;
    fn process_funds_withdrawals_for_execution(
        &self,
        chain_identifier: ChainIdentifier,
    ) -> BTreeMap<AccumulatorObjId, u64>;
    fn has_funds_withdrawals(&self) -> bool;
    fn get_funds_withdrawals(&self) -> Vec<FundsWithdrawalArg>;
    fn coin_reservation_obj_refs(
        &self,
        chain_identifier: ChainIdentifier,
    ) -> Vec<ParsedObjectRefWithdrawal>;
    fn validity_check(&self, context: &TxValidityCheckContext<'_>) -> SuiResult;
    fn validity_check_no_gas_check(
        &self,
        config: &ProtocolConfig,
    ) -> UserInputResult;
    fn check_sponsorship(&self) -> UserInputResult;
    fn is_system_tx(&self) -> bool;
    fn is_genesis_tx(&self) -> bool;
    fn is_end_of_epoch_tx(&self) -> bool;
    fn is_consensus_commit_prologue(&self) -> bool;
    fn is_sponsored_tx(&self) -> bool;
    fn is_gas_paid_from_address_balance(&self) -> bool;
    fn sender_mut_for_testing(&mut self) -> &mut SuiAddress;
    fn gas_data_mut(&mut self) -> &mut GasData;
    fn expiration_mut_for_testing(&mut self) -> &mut TransactionExpiration;

}

Required Methods

fn sender(&self) -> SuiAddress

fn kind(&self) -> &TransactionKind

fn kind_mut(&mut self) -> &mut TransactionKind

fn into_kind(self) -> TransactionKind

fn required_signers(&self) -> NonEmpty<SuiAddress>

Transaction signer and Gas owner

fn gas_data(&self) -> &GasData

fn gas_owner(&self) -> SuiAddress

fn gas(&self) -> &[ObjectRef]

fn gas_price(&self) -> u64

fn gas_budget(&self) -> u64

fn expiration(&self) -> &TransactionExpiration

fn expiration_mut(&mut self) -> &mut TransactionExpiration

fn move_calls(&self) -> Vec<(usize, &ObjectID, &str, &str)>

fn input_objects(&self) -> UserInputResult<Vec<InputObjectKind>>

fn shared_input_objects(&self) -> Vec<SharedInputObject>

fn receiving_objects(&self) -> Vec<ObjectRef>

fn fastpath_dependency_objects( &self, ) -> UserInputResult<(Vec<ObjectRef>, Vec<ObjectID>, Vec<ObjectRef>)>

fn process_funds_withdrawals_for_signing( &self, chain_identifier: ChainIdentifier, coin_resolver: &dyn CoinReservationResolverTrait, ) -> UserInputResult<BTreeMap<AccumulatorObjId, u64>>

Processes funds withdraws and returns a map from funds account object ID to total reserved amount. This method aggregates all withdraw operations for the same account by merging their reservations. Each account object ID is derived from the type parameter of each withdraw operation.

This method is used at signing time, and can reject a transaction if it contains invalid reservations.

fn process_funds_withdrawals_for_execution( &self, chain_identifier: ChainIdentifier, ) -> BTreeMap<AccumulatorObjId, u64>

Like process_funds_withdrawals_for_signing, but must only be called on a certified transaction, i.e. one that is known to be valid.

fn has_funds_withdrawals(&self) -> bool

fn get_funds_withdrawals(&self) -> Vec<FundsWithdrawalArg>

fn coin_reservation_obj_refs( &self, chain_identifier: ChainIdentifier, ) -> Vec<ParsedObjectRefWithdrawal>

fn validity_check(&self, context: &TxValidityCheckContext<'_>) -> SuiResult

fn validity_check_no_gas_check( &self, config: &ProtocolConfig, ) -> UserInputResult

fn check_sponsorship(&self) -> UserInputResult

Check if the transaction is compliant with sponsorship.

fn is_system_tx(&self) -> bool

fn is_genesis_tx(&self) -> bool

fn is_end_of_epoch_tx(&self) -> bool

returns true if the transaction is one that is specially sequenced to run at the very end of the epoch

fn is_consensus_commit_prologue(&self) -> bool

fn is_sponsored_tx(&self) -> bool

Check if the transaction is sponsored (namely gas owner != sender)

fn is_gas_paid_from_address_balance(&self) -> bool

fn sender_mut_for_testing(&mut self) -> &mut SuiAddress

fn gas_data_mut(&mut self) -> &mut GasData

fn expiration_mut_for_testing(&mut self) -> &mut TransactionExpiration