Trait sui_types::transaction::VersionedProtocolMessage

pub trait VersionedProtocolMessage {
    // Required method
    fn check_version_and_features_supported(
        &self,
        protocol_config: &ProtocolConfig
    ) -> SuiResult;

    // Provided method
    fn message_version(&self) -> Option<u64> { ... }
}

Required Methods

fn check_version_and_features_supported( &self, protocol_config: &ProtocolConfig ) -> SuiResult

Check that the version of the message is the correct one to use at this protocol version. Also checks whether the feauures used by the message are supported by the protocol config.

Provided Methods

fn message_version(&self) -> Option<u64>

Return version of message. Some messages depend on their enclosing messages to know the version number, so not every implementor implements this.