sui_types

Module transaction_driver_types

Structs

ExecuteTransactionRequestV3

ExecuteTransactionResponseV3

FinalizedEffects

TransactionDriverRequest

TransactionDriverResponse

Enums

EffectsFinalityInfo

Proof of finality of transaction effects.

ExecuteTransactionRequestType

ExecuteTransactionResponse

TransactionSubmissionError

Client facing errors regarding transaction submission via Transaction Driver. Every invariant needs detailed documents to instruct client handling.

TransactionType

Constants

NON_RECOVERABLE_ERROR_MSG

Type Aliases

GroupedErrors

IsTransactionExecutedLocally

When requested to execute a transaction with WaitForLocalExecution, TransactionOrchestrator attempts to execute this transaction locally after it is finalized. This value represents whether the transaction is confirmed to be executed on this node before the response returns.

TransactionDriverEffectsQueueResult

TransactionDriverResult