Crate sui_types

Modules

accumulator

authenticator_state

balance

base_types

clock

coin

collection_types

committee

crypto

digests

display

dynamic_field

effects

epoch_data

error

event

executable_transaction

execution

execution_mode

execution_status

full_checkpoint_content

gas

gas_coin

gas_model

governance

id

in_memory_storage

inner_temporary_store

message_envelope

messages_checkpoint

messages_consensus

messages_grpc

messages_safe_client

metrics

mock_checkpoint_builder

move_package

multiaddr

multisig

multisig_legacy

object

programmable_transaction_builder

Utility for generating programmable transactions, either by specifying a command or for migrating legacy transactions

quorum_driver_types

randomness_state

signature

storage

sui_serde

sui_system_state

transaction

transfer

type_resolver

utils

versioned

zk_login_authenticator

zk_login_util

Macros

assert_invariant

exit_main

fp_bail

fp_ensure

invariant_violation

make_invariant_violation

Structs

Identifier

An owned identifier.

Enums

TypeTag

Constants

DEEPBOOK_ADDRESS

0xdee9– account address where DeepBook modules are stored Same as the ObjectID

DEEPBOOK_PACKAGE_ID

MOVE_STDLIB_ADDRESS

0x1– account address where Move stdlib modules are stored Same as the ObjectID

MOVE_STDLIB_PACKAGE_ID

SUI_AUTHENTICATOR_STATE_ADDRESS

0x7: hardcode object ID for the singleton authenticator state object.

SUI_AUTHENTICATOR_STATE_OBJECT_ID

SUI_AUTHENTICATOR_STATE_OBJECT_SHARED_VERSION

SUI_CLOCK_ADDRESS

0x6: hardcoded object ID for the singleton clock object.

SUI_CLOCK_OBJECT_ID

SUI_CLOCK_OBJECT_SHARED_VERSION

SUI_FRAMEWORK_ADDRESS

0x2– account address where sui framework modules are stored Same as the ObjectID

SUI_FRAMEWORK_PACKAGE_ID

SUI_RANDOMNESS_STATE_ADDRESS

0x8: hardcode object ID for the singleton randomness state object.

SUI_RANDOMNESS_STATE_OBJECT_ID

SUI_SYSTEM_ADDRESS

0x3– account address where sui system modules are stored Same as the ObjectID

SUI_SYSTEM_PACKAGE_ID

SUI_SYSTEM_STATE_ADDRESS

0x5: hardcoded object ID for the singleton sui system state object.

SUI_SYSTEM_STATE_OBJECT_ID

SUI_SYSTEM_STATE_OBJECT_SHARED_VERSION

Traits

MoveTypeTagTrait

Functions

is_object

is_object_vector

is_primitive

is_system_package

Return true if addr is a special system package that can be upgraded at epoch boundaries. All new system package ID’s must be added here.

parse_sui_struct_tag

parse_sui_type_tag

sui_framework_address_concat_string