sui_types
Module bridge

Structs

BridgeCommitteeSummary

BridgeInnerV1

Rust version of the Move bridge::BridgeInner type.

BridgeLimiterSummary

BridgeSummary

BridgeTokenMetadata

BridgeTreasurySummary

BridgeWrapper

Rust version of the Move sui::bridge::Bridge type This repreents the object with 0x9 ID. In Rust, this type should be rarely used since it’s just a thin wrapper used to access the inner object. Within this module, we use it to determine the current version of the bridge inner object type, so that we could deserialize the inner object correctly.

MoveTypeBridgeCommittee

Rust version of the Move committee::BridgeCommittee type.

MoveTypeBridgeMessage

Rust version of the Move message::BridgeMessage type.

MoveTypeBridgeMessageKey

Rust version of the Move message::BridgeMessageKey type.

MoveTypeBridgeRecord

Rust version of the Move message::BridgeMessage type.

MoveTypeBridgeRoute

Rust version of the Move chain_ids::BridgeRoute type.

MoveTypeBridgeTransferLimiter

Rust version of the Move limiter::TransferLimiter type.

MoveTypeBridgeTransferRecord

Rust version of the Move limiter::TransferRecord type.

MoveTypeBridgeTreasury

Rust version of the Move treasury::BridgeTreasury type.

MoveTypeCommitteeMember

Rust version of the Move committee::CommitteeMember type.

MoveTypeCommitteeMemberRegistration

Rust version of the Move committee::CommitteeMemberRegistration type.

MoveTypeParsedTokenTransferMessage

Rust version of the Move message::ParsedTokenTransferMessage type.

MoveTypeTokenTransferPayload

Rust version of the Move message::TokenTransferPayload type.

Enums

Bridge

Bridge provides an abstraction over multiple versions of the inner BridgeInner object. This should be the primary interface to the bridge object in Rust. We use enum dispatch to dispatch all methods defined in BridgeTrait to the actual implementation in the inner types.

BridgeChainId

Constants

APPROVAL_THRESHOLD_ADD_TOKENS_ON_EVM

APPROVAL_THRESHOLD_ADD_TOKENS_ON_SUI

APPROVAL_THRESHOLD_ASSET_PRICE_UPDATE

APPROVAL_THRESHOLD_COMMITTEE_BLOCKLIST

APPROVAL_THRESHOLD_EMERGENCY_PAUSE

APPROVAL_THRESHOLD_EMERGENCY_UNPAUSE

APPROVAL_THRESHOLD_EVM_CONTRACT_UPGRADE

APPROVAL_THRESHOLD_LIMIT_UPDATE

APPROVAL_THRESHOLD_TOKEN_TRANSFER

BRIDGE_COMMITTEE_MAXIMAL_VOTING_POWER

BRIDGE_COMMITTEE_MINIMAL_VOTING_POWER

BRIDGE_COMMITTEE_MODULE_NAME

BRIDGE_CREATE_ADD_TOKEN_ON_SUI_MESSAGE_FUNCTION_NAME

BRIDGE_CREATE_FUNCTION_NAME

BRIDGE_EXECUTE_SYSTEM_MESSAGE_FUNCTION_NAME

BRIDGE_INIT_COMMITTEE_FUNCTION_NAME

BRIDGE_LIMITER_MODULE_NAME

BRIDGE_MESSAGE_MODULE_NAME

BRIDGE_MODULE_NAME

BRIDGE_REGISTER_FOREIGN_TOKEN_FUNCTION_NAME

BRIDGE_SUPPORTED_ASSET

BRIDGE_TREASURY_MODULE_NAME

TOKEN_ID_BTC

TOKEN_ID_ETH

TOKEN_ID_SUI

TOKEN_ID_USDC

TOKEN_ID_USDT

Traits

BridgeTrait

This is the standard API that all bridge inner object type should implement.

Functions

get_bridge

get_bridge_obj_initial_shared_version

get_bridge_wrapper

is_bridge_committee_initiated

Type Aliases

BridgeInnerDynamicField

BridgeRecordDyanmicField