sui_types::crypto
Trait Authenticator

pub trait Authenticator:
    ToFromBytes
    + Display
    + Serialize
    + DeserializeOwned
    + Send
    + Sync
    + 'static
    + Clone {
    type PubKey: VerifyingKey<Sig = Self>;
    type PrivKey: SigningKey<Sig = Self>;

    const LENGTH: usize;
}

Trait impl’d by signatures in asymmetric cryptography.

The trait bounds are implemented so as to be symmetric and equivalent to the ones on its associated types for private key and public key material.

Required Associated Constants

const LENGTH: usize

Required Associated Types

type PubKey: VerifyingKey<Sig = Self>

type PrivKey: SigningKey<Sig = Self>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.