Trait sui_types::crypto::AuthoritySignInfoTrait

pub trait AuthoritySignInfoTrait: SealedAuthoritySignInfoTrait {
    // Required methods
    fn verify_secure<T: Serialize>(
        &self,
        data: &T,
        intent: Intent,
        committee: &Committee
    ) -> SuiResult;
    fn add_to_verification_obligation<'a>(
        &self,
        committee: &'a Committee,
        obligation: &mut VerificationObligation<'a>,
        message_index: usize
    ) -> SuiResult<()>;
}

AuthoritySignInfoTrait is a trait used specifically for a few structs in messages.rs to template on whether the struct is signed by an authority. We want to limit how those structs can be instantiated on, hence the sealed trait. TODO: We could also add the aggregated signature as another impl of the trait. This will make CertifiedTransaction also an instance of the same struct.

Required Methods

fn verify_secure<T: Serialize>( &self, data: &T, intent: Intent, committee: &Committee ) -> SuiResult

fn add_to_verification_obligation<'a>( &self, committee: &'a Committee, obligation: &mut VerificationObligation<'a>, message_index: usize ) -> SuiResult<()>

Object Safety

This trait is not object safe.