sui_types
Module dynamic_field

Modules

visitor

Structs

DOFWrapper

Rust version of the Move sui::dynamic_object_field::Wrapper type

DynamicFieldInfo

DynamicFieldName

Field

Rust version of the Move sui::dynamic_field::Field type

Enums

DynamicFieldType

Constants

DYNAMIC_FIELD_FIELD_STRUCT_NAME

DYNAMIC_FIELD_MODULE_NAME

Functions

derive_dynamic_field_id

extract_field_from_move_struct

extract_id_value

get_dynamic_field_from_store

Similar to get_dynamic_field_object_from_store, but returns the value in the field instead of the Field object itself.

get_dynamic_field_object_from_store

Given a parent object ID (e.g. a table), and a key, retrieve the corresponding dynamic field object from the object_store. The key type K must implement MoveTypeTagTrait which has an associated function that returns the Move type tag. Note that this function returns the Field object itself, not the value in the field.

is_dynamic_object