sui_types
Module effects

Structs

AccumulatorAddress

Accumulator objects are named by an address (can be an account address or a UID) and a type tag.

AccumulatorWriteV1

EffectsObjectChange

ObjectChange

TestEffectsBuilder

TransactionEffectsDebugSummary

TransactionEffectsV1

The response from processing a transaction or a certified transaction

TransactionEffectsV2

The response from processing a transaction or a certified transaction

TransactionEvents

Enums

AccumulatorOperation

AccumulatorValue

IDOperation

InputSharedObject

ObjectIn

If an object exists (at root-level) in the store prior to this transaction, it should be Exist, otherwise it’s NonExist, e.g. wrapped objects should be NotExist.

ObjectOut

ObjectRemoveKind

TransactionEffects

The response from processing a transaction or a certified transaction

UnchangedSharedKind

Constants

APPROX_SIZE_OF_EPOCH_ID

APPROX_SIZE_OF_EXECUTION_STATUS

APPROX_SIZE_OF_GAS_COST_SUMMARY

APPROX_SIZE_OF_OBJECT_REF

APPROX_SIZE_OF_OPT_TX_EVENTS_DIGEST

APPROX_SIZE_OF_OWNER

APPROX_SIZE_OF_TX_DIGEST

Traits

TransactionEffectsAPI

Type Aliases

CertifiedTransactionEffects

SignedTransactionEffects

TransactionEffectsEnvelope

TrustedSignedTransactionEffects

UnsignedTransactionEffects

VerifiedCertifiedTransactionEffects

VerifiedSignedTransactionEffects

VerifiedTransactionEffectsEnvelope