sui_types::gas_model
Module gas_predicates

Functions

charge_input_as_memory

If true, input object bytes are treated as memory allocated in Move and charged according to the bucket they end up in.

charge_upgrades

cost_table_for_version

dont_charge_budget_on_storage_oog

If true, do not charge the entire budget on storage OOG

gas_price_too_high

If true, enable the check for gas price too high

native_function_threshold_exceeded

txn_base_cost_as_multiplier

use_legacy_abstract_size

If true, calculate value sizes using the legacy size calculation.