sui_types::gas_model::tables
Struct GasStatus

pub struct GasStatus {
    pub gas_model_version: u64,
    /* private fields */
}

The Move VM implementation of state for gas metering.

Initialize with a CostTable and the gas provided to the transaction. Provide all the proper guarantees about gas metering in the Move VM.

Every client must use an instance of this type to interact with the Move VM.

Fields
gas_model_version: u64
Implementations

impl GasStatus

pub fn new( cost_table: CostTable, budget: u64, gas_price: u64, gas_model_version: u64, ) -> Self

Initialize the gas state with metering enabled.

Charge for every operation and fail when there is no more gas to pay for operations. This is the instantiation that must be used when executing a user script.

pub fn new_unmetered() -> Self

Initialize the gas state with metering disabled.

It should be used by clients in very specific cases and when executing system code that does not have to charge the user.

pub fn push_stack(&mut self, pushes: u64) -> PartialVMResult<()>

pub fn pop_stack(&mut self, pops: u64)

pub fn increase_instruction_count(&mut self, amount: u64) -> PartialVMResult<()>

pub fn increase_stack_size(&mut self, size_amount: u64) -> PartialVMResult<()>

pub fn decrease_stack_size(&mut self, size_amount: u64)

pub fn charge( &mut self, num_instructions: u64, pushes: u64, pops: u64, incr_size: u64, _decr_size: u64, ) -> PartialVMResult<()>

Given: pushes + pops + increase + decrease in size for an instruction charge for the execution of the instruction.

pub fn cost_table(&self) -> &CostTable

Return the CostTable behind this GasStatus.

pub fn remaining_gas(&self) -> Gas

Return the gas left.

pub fn deduct_gas(&mut self, amount: InternalGas) -> PartialVMResult<()>

Charge a given amount of gas and fail if not enough gas units are left.

pub fn set_metering(&mut self, enabled: bool)

pub fn gas_used_pre_gas_price(&self) -> u64

pub fn charge_bytes( &mut self, size: usize, cost_per_byte: u64, ) -> PartialVMResult<()>

pub fn gas_price(&self) -> u64

pub fn stack_height_high_water_mark(&self) -> u64

pub fn stack_size_high_water_mark(&self) -> u64

pub fn instructions_executed(&self) -> u64