Struct sui_types::message_envelope::VerifiedEnvelope

pub struct VerifiedEnvelope<T: Message, S>(_, _);

Implementations

impl VerifiedEnvelope<SenderSignedData, CertificateProof>

pub fn certificate_sig(&self) -> Option<&AuthorityStrongQuorumSignInfo>

impl<T: Message, S> VerifiedEnvelope<T, S>

pub fn new_from_verified(inner: Envelope<T, S>) -> Self

This API should only be called when the input is already verified.

pub fn new_unchecked(inner: Envelope<T, S>) -> Self

There are some situations (e.g. fragment verification) where its very awkward and/or inefficient to obtain verified certificates from calling CertifiedTransaction::verify() Use this carefully.

pub fn into_inner(self) -> Envelope<T, S>

pub fn inner(&self) -> &Envelope<T, S>

pub fn into_message(self) -> T

pub fn serializable_ref(&self) -> &TrustedEnvelope<T, S>

Use this when you need to serialize a verified envelope. This should generally only be used for database writes. never use over the network!

pub fn serializable(self) -> TrustedEnvelope<T, S>

Use this when you need to serialize a verified envelope. This should generally only be used for database writes. never use over the network!

pub fn into_unsigned(self) -> VerifiedEnvelope<T, EmptySignInfo>

Remove the authority signatures S from this envelope.

impl<T: Message> VerifiedEnvelope<T, CertificateProof>

The following implementation provides two ways to construct a VerifiedEnvelope with CertificateProof. It is implemented in this file such that we could reuse the digest without having to recompute it. We allow converting a VerifiedCertificate into a VerifiedEnvelope with CertificateProof::Certificate; and converting a VerifiedTransaction along with checkpoint information into a VerifiedEnvelope with CertificateProof::Checkpoint.

pub fn new_from_certificate( certificate: VerifiedEnvelope<T, AuthorityStrongQuorumSignInfo> ) -> Self

pub fn new_from_checkpoint( transaction: VerifiedEnvelope<T, EmptySignInfo>, epoch: EpochId, checkpoint: CheckpointSequenceNumber ) -> Self

pub fn new_system( transaction: VerifiedEnvelope<T, EmptySignInfo>, epoch: EpochId ) -> Self

pub fn new_from_quorum_execution( transaction: VerifiedEnvelope<T, EmptySignInfo>, epoch: EpochId ) -> Self

pub fn epoch(&self) -> EpochId

impl VerifiedEnvelope<CheckpointSummary, AuthorityQuorumSignInfo<true>>

pub fn into_summary_and_sequence( self ) -> (CheckpointSequenceNumber, CheckpointSummary)

pub fn get_validator_signature(self) -> AggregateAuthoritySignature

impl VerifiedEnvelope<SenderSignedData, EmptySignInfo>

pub fn new_change_epoch( next_epoch: EpochId, protocol_version: ProtocolVersion, storage_charge: u64, computation_charge: u64, storage_rebate: u64, non_refundable_storage_fee: u64, epoch_start_timestamp_ms: u64, system_packages: Vec<(SequenceNumber, Vec<Vec<u8>>, Vec<ObjectID>)> ) -> Self

pub fn new_genesis_transaction(objects: Vec<GenesisObject>) -> Self

pub fn new_consensus_commit_prologue( epoch: u64, round: u64, commit_timestamp_ms: CheckpointTimestamp ) -> Self

impl VerifiedEnvelope<SenderSignedData, AuthoritySignInfo>

pub fn new( epoch: EpochId, transaction: VerifiedTransaction, authority: AuthorityName, secret: &dyn Signer<AuthoritySignature> ) -> Self

Use signing key to create a signed object.