Trait sui_types::message_envelope::Message

pub trait Message {
    type DigestType: Clone + Debug;

    fn digest(&self) -> Self::DigestType;
    fn verify(&self) -> SuiResult;
}

Required Associated Types

type DigestType: Clone + Debug

Required Methods

fn digest(&self) -> Self::DigestType

fn verify(&self) -> SuiResult

Verify the internal data consistency of this message. In some cases, such as user signed transaction, we also need to verify the user signature here.