Module sui_types::messages

Structs

AuthorityCapabilities

Used to advertise capabilities of each authority via narwhal. This allows validators to negotiate the creation of the ChangeEpoch transaction.

ChangeEpoch

ConsensusCommitPrologue

Only commit_timestamp_ms is passed to the move call currently. However we include epoch and round to make sure each ConsensusCommitPrologue has a unique tx digest.

ConsensusSync

ConsensusTransaction

ExecuteTransactionRequest

FinalizedEffects

GasData

GenesisTransaction

HandleCertificateResponse

HandleTransactionResponse

InputObjects

MoveLocation

MoveLocationOpt

ObjectInfoRequest

A request for information about an object and optionally its parent certificate at a specific version.

ObjectInfoResponse

This message provides information about the latest object and its lock.

ProgrammableMoveCall

The command for calling a Move function, either an entry function or a public function (which cannot return references).

ProgrammableTransaction

A series of commands where the results of one command can be used in future commands

QuorumDriverRequest

QuorumDriverResponse

SenderSignedData

SenderSignedTransaction

SharedInputObject

SubmitCertificateResponse

SystemStateRequest

TransactionDataV1

TransactionEffectsDebugSummary

TransactionEffectsV1

The response from processing a transaction or a certified transaction

TransactionEvents

TransactionInfoRequest

TransactionInfoResponse

VerifiedHandleCertificateResponse

VerifiedObjectInfoResponse

Verified version of ObjectInfoResponse. layout and lock_for_debugging are skipped because they are not needed and we don’t want to verify them.

Enums

Argument

An argument to a programmable transaction command

CallArg

Command

A single command in a programmable transaction.

CommandArgumentError

ConsensusTransactionKey

ConsensusTransactionKind

EffectsFinalityInfo

ExecuteTransactionRequestType

ExecuteTransactionResponse

ExecutionFailureStatus

ExecutionStatus

GenesisObject

InputObjectKind

ObjectArg

ObjectInfoRequestKind

PackageUpgradeError

PlainTransactionInfoResponse

This enum represents all possible states of a response returned from the safe client. Note that [struct SignedTransaction] and [struct SignedTransactionEffects] are represented as an Envelope instead of an VerifiedEnvelope. This is because the verification is now performed by the authority aggregator as an aggregated signature, instead of in SafeClient.

TransactionData

TransactionEffects

The response from processing a transaction or a certified transaction

TransactionExpiration

TransactionKind

TransactionStatus

TransactionType

TypeArgumentError

Constants

APPROX_SIZE_OF_EPOCH_ID

APPROX_SIZE_OF_EXECUTION_STATUS

APPROX_SIZE_OF_GAS_COST_SUMMARY

APPROX_SIZE_OF_OBJECT_REF

APPROX_SIZE_OF_OPT_TX_EVENTS_DIGEST

APPROX_SIZE_OF_OWNER

APPROX_SIZE_OF_TX_DIGEST

GAS_PRICE_FOR_SYSTEM_TX

TEST_ONLY_GAS_UNIT_FOR_GENERIC

TEST_ONLY_GAS_UNIT_FOR_OBJECT_BASICS

TEST_ONLY_GAS_UNIT_FOR_PUBLISH

TEST_ONLY_GAS_UNIT_FOR_SPLIT_COIN

TEST_ONLY_GAS_UNIT_FOR_STAKING

TEST_ONLY_GAS_UNIT_FOR_TRANSFER

TEST_ONLY_GAS_UNIT_FOR_VALIDATOR

Traits

TransactionDataAPI

TransactionEffectsAPI

VersionedProtocolMessage

Type Definitions

CertifiedTransaction

CertifiedTransactionEffects

CommandIndex

ExecutableTransaction

An ExecutableTransaction is a wrapper of a transaction with a CertificateProof that indicates there existed a valid certificate for this transaction, and hence it can be executed locally. This is an abstraction data structure to cover both the case where the transaction is certified or checkpointed when we schedule it for execution.

IsTransactionExecutedLocally

When requested to execute a transaction with WaitForLocalExecution, TransactionOrchestrator attempts to execute this transaction locally after it is finalized. This value represents whether the transaction is confirmed to be executed on this node before the response returns.

SignedTransaction

A transaction that is signed by a sender and also by an authority.

SignedTransactionEffects

Transaction

A transaction that is signed by a sender but not yet by an authority.

TransactionEffectsEnvelope

TrustedCertificate

TrustedExecutableTransaction

TrustedSignedTransactionEffects

TrustedTransaction

TxCertAndSignedEffects

UnsignedTransactionEffects

VerifiedCertificate

VerifiedCertifiedTransactionEffects

VerifiedExecutableTransaction

VerifiedSignedTransaction

VerifiedSignedTransactionEffects

VerifiedTransaction

VerifiedTransactionEffectsEnvelope