Module sui_types::messages

Structs

AuthorityCapabilities

Used to advertise capabilities of each authority via narwhal. This allows validators to negotiate the creation of the ChangeEpoch transaction.

ChangeEpoch

CircularObjectOwnership

ConsensusCommitPrologue

Only commit_timestamp_ms is passed to the move call currently. However we include epoch and round to make sure each ConsensusCommitPrologue has a unique tx digest.

ConsensusSync

ConsensusTransaction

EntryArgumentError

EntryTypeArgumentError

ExecuteTransactionRequest

FinalizedEffects

GasData

GenesisTransaction

HandleCertificateResponse

HandleTransactionResponse

InputObjects

InvalidChildObjectArgument

InvalidSharedByValue

MoveCall

MoveLocation

MoveModulePublish

ObjectInfoRequest

A request for information about an object and optionally its parent certificate at a specific version.

ObjectInfoResponse

This message provides information about the latest object and its lock.

Pay

Pay each recipient the corresponding amount using the input coins

PayAllSui

Send all SUI coins to one recipient. only for SUI coin and does not require a separate gas coin object either. Specifically, what pay_all_sui does are:

PaySui

Send SUI coins to a list of addresses, following a list of amounts. only for SUI coin and does not require a separate gas coin object. Specifically, what pay_sui does are:

ProgrammableMoveCall

The command for calling a Move function, either an entry function or a public function (which cannot return references).

ProgrammableTransaction

A series of commands where the results of one command can be used in future commands

QuorumDriverRequest

QuorumDriverResponse

SenderSignedData

SharedInputObject

SystemStateRequest

TransactionDataV1

TransactionEffectsDebugSummary

TransactionEffectsV1

The response from processing a transaction or a certified transaction

TransactionEvents

TransactionInfoRequest

TransactionInfoResponse

TransferObject

TransferSui

VerifiedHandleCertificateResponse

VerifiedObjectInfoResponse

Verified version of ObjectInfoResponse. layout and lock_for_debugging are skipped because they are not needed and we don’t want to verify them.

Enums

Argument

An argument to a programmable transaction command

CallArg

CallResult

Command

A single command in a programmable transaction.

CommandArgumentError

ConsensusTransactionKey

ConsensusTransactionKind

EffectsFinalityInfo

EntryArgumentErrorKind

EntryTypeArgumentErrorKind

ExecuteTransactionRequestType

ExecuteTransactionResponse

ExecutionFailureStatus

ExecutionStatus

GenesisObject

InputObjectKind

ObjectArg

ObjectInfoRequestKind

SingleTransactionKind

TransactionData

TransactionEffects

The response from processing a transaction or a certified transaction

TransactionExpiration

TransactionKind

TransactionStatus

TransactionType

VerifiedTransactionInfoResponse

Constants

DUMMY_GAS_PRICE

Traits

TransactionDataAPI

TransactionEffectsAPI

VersionedProtocolMessage

Type Definitions

CertifiedTransaction

CertifiedTransactionEffects

CommandIndex

ExecutableTransaction

An ExecutableTransaction is a wrapper of a transaction with a CertificateProof that indicates there existed a valid certificate for this transaction, and hence it can be executed locally. This is an abstraction data structure to cover both the case where the transaction is certified or checkpointed when we schedule it for execution.

IsTransactionExecutedLocally

When requested to execute a transaction with WaitForLocalExecution, TransactionOrchestrator attempts to execute this transaction locally after it is finalized. This value represents whether the transaction is confirmed to be executed on this node before the response returns.

SignedTransaction

A transaction that is signed by a sender and also by an authority.

SignedTransactionEffects

Transaction

A transaction that is signed by a sender but not yet by an authority.

TransactionEffectsEnvelope

TrustedCertificate

TrustedExecutableTransaction

TrustedSignedTransactionEffects

TrustedTransaction

TxCertAndSignedEffects

UnsignedTransactionEffects

VerifiedCertificate

VerifiedCertifiedTransactionEffects

VerifiedExecutableTransaction

VerifiedSignedTransaction

VerifiedSignedTransactionEffects

VerifiedTransaction

VerifiedTransactionEffectsEnvelope