Module sui_types::messages

Structs

AccountInfoRequest

AccountInfoResponse

ChangeEpoch

CircularObjectOwnership

CommitteeInfo

CommitteeInfoRequest

CommitteeInfoResponse

ConsensusSync

ConsensusTransaction

EntryArgumentError

EntryTypeArgumentError

ExecuteTransactionRequest

GenesisTransaction

HandleCertificateResponse

InputObjects

InvalidChildObjectArgument

InvalidSharedByValue

MoveCall

MoveLocation

MoveModulePublish

ObjectInfoRequest

A request for information about an object and optionally its parent certificate at a specific version.

ObjectInfoResponse

This message provides information about the latest object and its lock as well as the parent certificate of the object at a specific version.

ObjectResponse

Pay

Pay each recipient the corresponding amount using the input coins

PayAllSui

Send all SUI coins to one recipient. only for SUI coin and does not require a separate gas coin object either. Specifically, what pay_all_sui does are:

PaySui

Send SUI coins to a list of addresses, following a list of amounts. only for SUI coin and does not require a separate gas coin object. Specifically, what pay_sui does are:

QuorumDriverRequest

QuorumDriverResponse

SenderSignedData

SharedInputObject

TransactionData

TransactionEffects

The response from processing a transaction or a certified transaction

TransactionInfoRequest

TransactionInfoResponse

TransferObject

TransferSui

VerifiedHandleCertificateResponse

Enums

CallArg

CallResult

ConsensusTransactionKey

ConsensusTransactionKind

EntryArgumentErrorKind

EntryTypeArgumentErrorKind

ExecuteTransactionRequestType

ExecuteTransactionResponse

ExecutionFailureStatus

ExecutionStatus

GenesisObject

InputObjectKind

ObjectArg

ObjectInfoRequestKind

SingleTransactionKind

TransactionKind

Constants

DUMMY_GAS_PRICE

Type Definitions

CertifiedTransaction

CertifiedTransactionEffects

CommitteeInfoResponseDigest

IsTransactionExecutedLocally

When requested to execute a transaction with WaitForLocalExecution, TransactionOrchestrator attempts to execute this transaction locally after it is finalized. This value represents whether the transaction is confirmed to be executed on this node before the response returns.

SignedTransaction

A transaction that is signed by a sender and also by an authority.

SignedTransactionEffects

Transaction

A transaction that is signed by a sender but not yet by an authority.

TransactionEffectsEnvelope

TrustedCertificate

TrustedSignedTransactionEffects

TxCertAndSignedEffects

UnsignedTransactionEffects

ValidExecutionDigests

ValidTransactionEffects

ValidTransactionEffectsDigest

VerifiedCertificate

VerifiedCertifiedTransactionEffects

VerifiedObjectInfoResponse

VerifiedSignedTransaction

VerifiedSignedTransactionEffects

VerifiedTransaction

VerifiedTransactionEffectsEnvelope

VerifiedTransactionInfoResponse