Module sui_types::messages_checkpoint

Re-exports

pub use crate::digests::CheckpointContentsDigest;

pub use crate::digests::CheckpointDigest;

Structs

CheckpointContentsV1

CheckpointContents are the transactions included in an upcoming checkpoint. They must have already been causally ordered. Since the causal order algorithm is the same among validators, we expect all honest validators to come up with the same order for each checkpoint content.

CheckpointRequest

CheckpointRequestV2

CheckpointResponse

CheckpointResponseV2

CheckpointSignatureMessage

This is a message validators publish to consensus in order to sign checkpoint

CheckpointSummary

CheckpointVersionSpecificDataV1

ECMHLiveObjectSetDigest

The Sha256 digest of an EllipticCurveMultisetHash committing to the live object set.

EndOfEpochData

FullCheckpointContents

Same as CheckpointContents, but contains full contents of all Transactions and TransactionEffects associated with the checkpoint.

VerifiedCheckpointContents

Enums

CheckpointCommitment

CheckpointContents

CheckpointSummaryResponse

CheckpointVersionSpecificData

Holds data in CheckpointSummary that is serialized into the version_specific_data field.

Type Aliases

CertifiedCheckpointSummary

CheckpointSequenceNumber

CheckpointSummaryEnvelope

CheckpointTimestamp

SignedCheckpointSummary

TrustedCheckpoint

VerifiedCheckpoint