sui_types
Module messages_consensus

Structs

AuthorityCapabilitiesV1

Used to advertise capabilities of each authority via consensus. This allows validators to negotiate the creation of the ChangeEpoch transaction.

AuthorityCapabilitiesV2

Used to advertise capabilities of each authority via consensus. This allows validators to negotiate the creation of the ChangeEpoch transaction.

ConsensusCommitPrologue

Only commit_timestamp_ms is passed to the move call currently. However we include epoch and round to make sure each ConsensusCommitPrologue has a unique tx digest.

ConsensusCommitPrologueV2

ConsensusCommitPrologueV3

ConsensusCommitPrologueV4

ConsensusPosition

The position of a transaction in consensus.

ConsensusTransaction

ExecutionTimeObservation

Used to share estimates of transaction execution times with other validators for congestion control.

Enums

ConsensusDeterminedVersionAssignments

Uses an enum to allow for future expansion of the ConsensusDeterminedVersionAssignments.

ConsensusTransactionKey

ConsensusTransactionKind

VersionedDkgConfirmation

VersionedDkgMessage

Functions

check_total_jwk_size

Type Aliases

AuthorityIndex

The index of an authority in the consensus committee. The value should be the same in Sui committee.

Round

Consensus round number in u64 instead of u32 for compatibility with Narwhal.

TimestampMs

Non-decreasing timestamp produced by consensus in ms.

TransactionIndex

The index of a transaction in a consensus block.