sui_types
Module messages_grpc

Structs

HandleCertificateRequestV3

HandleCertificateResponseV2

HandleCertificateResponseV3

Response type for version 3 of the handle certificate validator API.

HandleSoftBundleCertificatesRequestV3

Soft Bundle request. See SIP-19.

HandleSoftBundleCertificatesResponseV3

Response type for the handle Soft Bundle certificates validator API. If wait_for_effects is true, it is guaranteed that:

HandleTransactionResponse

ObjectInfoRequest

A request for information about an object and optionally its parent certificate at a specific version.

ObjectInfoResponse

This message provides information about the latest object and its lock.

RawExecutedData

RawExecutedStatus

RawRejectedStatus

RawSubmitTxRequest

RawSubmitTxResponse

RawWaitForEffectsRequest

RawWaitForEffectsResponse

SubmitCertificateResponse

SubmitTxResponse

SystemStateRequest

TransactionInfoRequest

TransactionInfoResponse

VerifiedHandleCertificateResponse

VerifiedObjectInfoResponse

Verified version of ObjectInfoResponse. layout and lock_for_debugging are skipped because they are not needed and we don’t want to verify them.

Enums

LayoutGenerationOption

Layout generation options – you can either generate or not generate the layout.

ObjectInfoRequestKind

RawRejectReason

RawValidatorTransactionStatus

TransactionStatus