sui_types
Module move_package

Structs

FnInfo

Additional information about a function

FnInfoKey

Uniquely identifies a function in a module

MovePackage

TypeOrigin

Identifies a struct and the module it was defined in

UpgradeCap

Rust representation of sui::package::UpgradeCap.

UpgradeInfo

Upgraded package info for the linkage table

UpgradeReceipt

Rust representation of sui::package::UpgradeReceipt.

UpgradeTicket

Rust representation of sui::package::UpgradeTicket.

Enums

UpgradePolicy

Rust representation of upgrade policy constants in sui::package.

Constants

PACKAGE_MODULE_NAME

UPGRADECAP_STRUCT_NAME

UPGRADERECEIPT_STRUCT_NAME

UPGRADETICKET_STRUCT_NAME

Statics

MAX_DISASSEMBLED_MODULE_SIZE

Functions

is_test_fun

Checks if a function is annotated with one of the test-related annotations

normalize_deserialized_modules

If include_code is set to false, the normalized module will skip function bodies but still include the signatures.

normalize_modules

If include_code is set to false, the normalized module will skip function bodies but still include the signatures.

Type Aliases

FnInfoMap

A map from function info keys to function info