Module sui_types::multisig

Structs

MultiSig

The struct that contains signatures and public keys necessary for authenticating a MultiSig.

MultiSigPublicKey

The struct that contains the public key used for authenticating a MultiSig.

Constants

MAX_BITMAP_VALUE

MAX_SIGNER_IN_MULTISIG

Functions

as_indices

Interpret a bitmap of 01s as a list of indices that is set to 1s. e.g. 22 = 0b10110, then the result is [1, 2, 4].

Type Aliases

BitmapUnit

ThresholdUnit

WeightUnit

Attribute Macros

enum_dispatch

Annotating a trait or enum definition with an #[enum_dispatch] attribute will register it with the enum_dispatch library, allowing it to be used to generate impl blocks elsewhere.