sui_types
Module multisig_legacy

Structs

MultiSigLegacy

Deprecated, use [struct MultiSig] instead. The struct that contains signatures and public keys necessary for authenticating a MultiSigLegacy.

MultiSigPublicKeyLegacy

Deprecated, use [struct MultiSigPublicKey] instead. The struct that contains the public key used for authenticating a MultiSig.

Constants

MAX_SIGNER_IN_MULTISIG

Functions

bitmap_to_u16

Convert a roaring bitmap to plain bitmap.

Type Aliases

ThresholdUnit

WeightUnit

Attribute Macros

enum_dispatch

Annotating a trait or enum definition with an #[enum_dispatch] attribute will register it with the enum_dispatch library, allowing it to be used to generate impl blocks elsewhere.