Module sui_types::quorum_driver_types

Structs

ExecuteTransactionRequest

FinalizedEffects

QuorumDriverRequest

QuorumDriverResponse

Enums

EffectsFinalityInfo

ExecuteTransactionRequestType

ExecuteTransactionResponse

QuorumDriverError

Client facing errors regarding transaction submission via Quorum Driver. Every invariant needs detailed documents to instruct client handling.

TransactionType

Constants

NON_RECOVERABLE_ERROR_MSG

Type Definitions

GroupedErrors

IsTransactionExecutedLocally

When requested to execute a transaction with WaitForLocalExecution, TransactionOrchestrator attempts to execute this transaction locally after it is finalized. This value represents whether the transaction is confirmed to be executed on this node before the response returns.

QuorumDriverEffectsQueueResult

QuorumDriverResult