Module sui_types::storage

Modules

error

Structs

ObjectKey

PackageObject

SharedInMemoryStore

SingleCheckpointSharedInMemoryStore

Enums

DeleteKind

DeleteKindWithOldVersion

DeleteKind together with the old sequence number prior to the deletion, if available. For normal deletion and wrap, we always will consult the object store to obtain the old sequence number. For UnwrapThenDelete however, in the old protocol where simplified_unwrap_then_delete is false, we will consult the object store to obtain the old sequence number, which latter will be put in modified_at_versions; in the new protocol where simplified_unwrap_then_delete is true, we will not consult the object store, and hence won’t have the old sequence number.

InputKey

A potential input to a transaction.

MarkerValue

ObjectChange

ObjectOrTombstone

WriteKind

Traits

BackingPackageStore

BackingStore

ChildObjectResolver

An abstraction of the (possibly distributed) store for objects. This API only allows for the retrieval of objects, not any state changes

GetSharedLocks

ObjectStore

ParentSync

ReadStore

Storage

An abstraction of the (possibly distributed) store for objects, and (soon) events and transactions

StorageView

WriteStore

Functions

get_module

get_module_by_id

get_package_objects

Returns Ok(<package object for each package id in package_ids>) if all package IDs in package_id were found. If any package in package_ids was not found it returns a list of any package ids that are unable to be found>).

load_package_object_from_object_store

transaction_input_object_keys

Fetch the ObjectKeys (IDs and versions) for non-shared input objects. Includes owned, and immutable objects as well as the gas objects, but not move packages or shared objects.

transaction_receiving_object_keys

Type Aliases

PackageFetchResults