sui_types::storage
Trait RpcStateReader

pub trait RpcStateReader:
    ObjectStore
    + ReadStore
    + Send
    + Sync {
    // Required methods
    fn get_lowest_available_checkpoint_objects(
        &self,
    ) -> Result<CheckpointSequenceNumber>;
    fn get_chain_identifier(&self) -> Result<ChainIdentifier>;
    fn indexes(&self) -> Option<&dyn RpcIndexes>;
    fn get_struct_layout(
        &self,
        type_tag: &StructTag,
    ) -> Result<Option<MoveTypeLayout>>;

    // Provided method
    fn get_type_layout(
        &self,
        type_tag: &TypeTag,
    ) -> Result<Option<MoveTypeLayout>> { ... }
}

Trait used to provide functionality to the REST API service.

It extends both ObjectStore and ReadStore by adding functionality that may require more detailed underlying databases or indexes to support.

Required Methods

fn get_lowest_available_checkpoint_objects( &self, ) -> Result<CheckpointSequenceNumber>

Lowest available checkpoint for which object data can be requested.

Specifically this is the lowest checkpoint for which input/output object data will be available.

fn get_chain_identifier(&self) -> Result<ChainIdentifier>

fn indexes(&self) -> Option<&dyn RpcIndexes>

fn get_struct_layout( &self, type_tag: &StructTag, ) -> Result<Option<MoveTypeLayout>>

Provided Methods

fn get_type_layout(&self, type_tag: &TypeTag) -> Result<Option<MoveTypeLayout>>