sui_types::sui_system_state

Module sui_system_state_summary

Structs

SuiSystemStateSummary

This is the JSON-RPC type for the SUI system state object. It flattens all fields to make them top-level fields such that it as minimum dependencies to the internal data structures of the SUI system state type.

SuiValidatorSummary

This is the JSON-RPC type for the SUI validator. It flattens all inner structures to top-level fields so that they are decoupled from the internal definitions.

Functions

get_validator_by_pool_id

Given the staking pool id of a validator, return the validator’s SuiValidatorSummary, works for validator candidates, active validators, as well as inactive validators.