sui_types
Module traffic_control

Structs

FreqThresholdConfig

PolicyConfig

RemoteFirewallConfig

TrafficControlReconfigParams

Weight

Enums

ClientIdSource

The source that should be used to identify the client’s IP address. To be used to configure cases where a node has infra running in front of the node that is separate from the protocol, such as a load balancer. Note that this is not the same as the client type (e.g a direct client vs a proxy client, as in the case of a fullnode driving requests from many clients).

PolicyType

Constants

DEFAULT_SKETCH_CAPACITY

DEFAULT_SKETCH_PROBABILITY

DEFAULT_SKETCH_TOLERANCE

Functions

default_channel_capacity

default_client_id_source

default_connection_blocklist_ttl_sec

default_dry_run

default_spam_sample_rate