Module sui_types::transaction

Structs

AuthenticatorStateExpire

AuthenticatorStateUpdate

ChangeEpoch

CheckedInputObjects

GasData

GenesisTransaction

InputObjects

ObjectReadResult

The result of reading an object for execution. Because shared objects may be deleted, one possible result of reading a shared object is that ObjectReadResultKind::Deleted is returned.

ProgrammableMoveCall

The command for calling a Move function, either an entry function or a public function (which cannot return references).

ProgrammableTransaction

A series of commands where the results of one command can be used in future commands

RandomnessStateUpdate

ReceivingObjectReadResult

ReceivingObjects

SenderSignedData

SenderSignedTransaction

SharedInputObject

TransactionDataV1

Enums

Argument

An argument to a programmable transaction command

CallArg

Command

A single command in a programmable transaction.

EndOfEpochTransactionKind

EndOfEpochTransactionKind

GenesisObject

InputObjectKind

ObjectArg

ObjectReadResultKind

ReceivingObjectReadResultKind

TransactionData

TransactionExpiration

TransactionKey

TransactionKey uniquely identifies a transaction across all epochs. Note that a single transaction may have multiple keys, for example a RandomnessStateUpdate could be identified by both Digest and RandomnessRound.

TransactionKind

Constants

DEFAULT_VALIDATOR_GAS_PRICE

GAS_PRICE_FOR_SYSTEM_TX

TEST_ONLY_GAS_UNIT_FOR_GENERIC

TEST_ONLY_GAS_UNIT_FOR_HEAVY_COMPUTATION_STORAGE

TEST_ONLY_GAS_UNIT_FOR_OBJECT_BASICS

TEST_ONLY_GAS_UNIT_FOR_PUBLISH

TEST_ONLY_GAS_UNIT_FOR_SPLIT_COIN

TEST_ONLY_GAS_UNIT_FOR_STAKING

TEST_ONLY_GAS_UNIT_FOR_TRANSFER

Traits

TransactionDataAPI

Functions

write_sep

Type Aliases

CertifiedTransaction

SignedTransaction

A transaction that is signed by a sender and also by an authority.

Transaction

A transaction that is signed by a sender but not yet by an authority.

TrustedCertificate

TrustedTransaction

VerifiedCertificate

VerifiedSignedTransaction

VerifiedTransaction